==> code/04_diversity_and_richness.Rmd <==
---
title: "Small Mammal Diversity"
author: "Kate Culhane"
output: html_document
editor_options: 
  chunk_output_type: console
---

```{r setup, include=FALSE}
knitr::opts_knit$set(root.dir = '..')
knitr::opts_chunk$set(echo = FALSE, warning = FALSE, message = FALSE,
                      fig.width=5, fig.height=3)
```

<br>

```{r packages and data}

# Load packages

library(tidyverse)
library(vegan)
library(VennDiagram)
library(iNEXT)
library(dunn.test)
library(rstatix)
library(ggpubr)
library(effectsize)

# Load data

matrix_df <- read_csv('output_data/species_matrix.csv') %>% 
  mutate(severity = factor(severity, levels = c('unb', 'mod', 'high')))

matrix_adj_df <- read_csv('output_data/species_matrix_adj.csv') %>% 
  mutate(severity = factor(severity, levels = c('unb', 'mod', 'high')))

effort <- read_csv('output_data/effort.csv')

```

```{r data wrangling}

# Transform data frames into matrices

matrix <- as.matrix(matrix_df[-c(1,2)])
rownames(matrix) <- matrix_df$site

matrix_adj <- as.matrix(matrix_adj_df[-c(1,2)])
rownames(matrix_adj) <- matrix_adj_df$site

# Create separate species matrices for each treatment

matrix_unb <- matrix[1:9,] # species matrix for low severity
matrix_mod <- matrix[10:18,] # species matrix for med severity
matrix_high <- matrix[19:27,] # species matrix for high severity

# Create matrix with species pooled for each treatment

matrix_pooled_df <- matrix_df %>% 
  group_by(severity) %>% 
  summarize_if(is.numeric, sum)

matrix_pooled <- as.matrix(matrix_pooled_df[-1])
rownames(matrix_pooled) <- matrix_pooled_df$severity

# Vectors for boxplots

cols <- c('darkgreen', 'orange', 'red')
xlabs <- c('Unburned', 'Moderate Severity', 'High Severity') 

```

### Species Richness

#### Number of species

**Number of species per site, by fire severity:**

```{r}

# Calculate species richness (number of species) for each site
n <- specnumber(matrix)

# Make a dataframe with site, severity, and number of species
n_spp <- tibble(site = matrix_df$site, severity = matrix_df$severity, n_spp = n)

# Jitterplot with mean/SE
ggplot(n_spp, aes(x = severity, y = n)) +
  geom_jitter(aes(color = severity)) +
  scale_color_manual(values = cols) +
  stat_summary(fun.data = mean_se, geom = "crossbar", width = .5) +
  theme_classic() +
  labs(x = 'Fire Severity',
       y = 'Number of Species') +
  theme(legend.position = 'NA') +
  scale_x_discrete(labels = xlabs)

```

<br>

**Venn diagram of the number of species found in each fire severity:** 

```{r species Venn diagram, fig.width=2, fig.height=2, echo=FALSE, eval=FALSE}

# Vectors of species at each severity

df <- gather(matrix_pooled_df, key = species, value = n, 2:12) %>% 
  filter(n != 0)

unb <- filter(df, severity == 'unb')
mod <- filter(df, severity == 'mod')
high <- filter(df, severity == 'high')

all_spp <- list(unb$species, mod$species, high$species)
names(all_spp) <- c('Unburned', 'Low-Moderate', 'High')

# VennDiagram
spp_venn <- venn.diagram(all_spp,
                         filename = NULL,
                         lty = 'blank',
                         cex = 1,
                         fontfamily = 'sans',
                         cat.pos = c(-18, 18, 180),
                         cat.dist = c(0.07, 0.07, 0.05),
                         cat.cex = 0.9,
                         cat.fontfamily = 'sans',
                         fill = c('darkgreen', 'orange', 'red'))
plot.new()
grid.draw(spp_venn)

```

- Nine species were found at unburned sites, eight at low-moderate severity sites, and seven at high severity sites
- Shared species: *P. maniculatus*, *N. quadrimaculatus*, *O. beecheyi*, *P. boylii*
- Species unique to unburned sites: *G. sabrinus*, *R. megalotis*

<br>

#### Individual-based rarefaction

Sites pooled by severity

```{r rarefaction curves, fig.height=4}

# Rarefaction curves

accum_unb <- specaccum(matrix_unb, method = 'rarefaction', permutations = 1000)
accum_mod <- specaccum(matrix_mod, method = 'rarefaction', permutations = 1000)
accum_high <- specaccum(matrix_high, method = 'rarefaction', permutations = 1000)

# Plot curves

plot(accum_unb,
     ci = 2, # 95% confidence interval
     ci.type = 'polygon', ci.lty = 0,
     ci.col = rgb(red=0, green=1, blue=0, alpha=0.25),
     lwd = 2, col = 'darkgreen')
lines(accum_mod,
      ci = 2, # 95% confidence interval
      ci.type = 'polygon',
      ci.col = rgb(red=1, green=0.5, blue=0, alpha=0.25), ci.lty = 0,
      lwd = 2, col = 'orange')
lines(accum_high,
      ci = 2, # 95% confidence interval
      ci.type = 'polygon', ci.lty = 0,
      ci.col = rgb(red=1, green=0, blue=0, alpha=0.25),
      lwd = 2, col = 'red')

```

<br>

#### Individual-based rarefaction, scaled to individuals

```{r}

# ggplot using individual-based rarefaction

unb_data <- data.frame(site = accum_unb$sites,
                       indv = accum_unb$individuals,
                       richness = accum_unb$richness,
                       CI = accum_unb$sd)
mod_data <- data.frame(site = accum_mod$sites,
                       indv = accum_mod$individuals,
                       richness = accum_mod$richness,
                       CI = accum_mod$sd)
high_data <- data.frame(site = accum_high$sites,
                        indv = accum_high$individuals,
                        richness = accum_high$richness,
                        CI = accum_high$sd)

ggplot() +
  geom_line(data = unb_data,
            aes(x = indv, y = richness),
            size = 1, color = 'darkgreen') +
  geom_ribbon(data = unb_data,
              aes(x = indv, ymin = (richness-2*CI), ymax=(richness+2*CI)),
              alpha = 0.2, fill = 'darkgreen') +
  geom_line(data = mod_data,
            aes(x = indv, y = richness),
            size = 1, color = 'orange') +  
  geom_ribbon(data = mod_data,
              aes(x = indv, ymin = (richness-2*CI), ymax=(richness+2*CI)),
              alpha = 0.2, fill = 'orange') +
  geom_line(data = high_data,
            aes(x = indv, y = richness),
            size = 1, color = 'red') +
  geom_ribbon(data = high_data,
              aes(x = indv, ymin = (richness-2*CI), ymax=(richness+2*CI)),
              alpha = 0.2, fill = 'red') +
  scale_y_continuous(breaks = c(2,4,6,8,10)) +
  labs(x = 'Individuals Caught',
       y = 'Richness') +
  theme_classic() +
  theme(axis.title.x = element_text(size = 13),
        axis.title.y = element_text(size = 13),
        axis.text.x = element_text(size = 11, color = 'black'),
        axis.text.y = element_text(size = 11, color = 'black'),
        plot.margin = unit(c(10,10,10,10), 'pt'))

```

```{r legend, eval=FALSE}

# Stuff for legend

x <- data.frame(letter = c('a','a','b','b','c','c'),
                value = c(1,1,2,2,3,3))

ggplot(x, aes(x = value, y = value, color = letter)) +
  geom_line(size = 1.3) +
  geom_ribbon(aes(ymin = value+0.5, ymax = value-0.5, fill = letter), color = NA, alpha = 0.2) +
  scale_color_manual(values = c('darkgreen', 'orange', 'red')) +
  scale_fill_manual(values = c('darkgreen', 'orange', 'red')) +
  theme(legend.key.size = unit(1, 'cm'))

```

<br>

#### Individual-based rarefaction with unconstrained 95% confidence intervals

```{r, fig.width=7}

# Individual based rarefaction

inext_matrix <- t(matrix_pooled)

out <- iNEXT(inext_matrix, q = 0, datatype = 'abundance')

# Plot curves

df <- fortify(out, type=1) %>% 
  mutate(site = factor(site, levels = c('unb', 'mod', 'high')))

df <- fortify(out, type=1) %>% 
  mutate(site = factor(site, levels = c('unb', 'mod', 'high')))
df.point <- df[which(df$method=="observed"),]
df.line <- df[which(df$method!="observed"),]
df.line$method <- factor(df.line$method, 
                         c("interpolated", "extrapolated"),
                         c("interpolation", "extrapolation"))
 
ggplot(df, aes(x = x, y = y, color = site)) + 
  geom_line(data = df.line, aes(linetype = method),
            lwd = 1) +
  geom_ribbon(aes(ymin = y.lwr, ymax = y.upr, fill = site, color = NULL),
              alpha=0.2) +
  scale_linetype(name = 'Method', labels = c('Interpolation', 'Extrapolation')) +
  scale_color_manual(name = 'Severity', values = c('#4AB793', '#DCA827', '#D85F2B'),
                     labels = c('Unburned', 'Low-Moderate Severity', 'High Severity')) +
  scale_fill_manual(name = 'Severity', values = c('darkgreen', 'orange', 'red'),
                    labels = c('Unburned', 'Low-Moderate Severity', 'High Severity')) +
  labs(x = 'Number of Individuals',
       y = 'Rarefied Richness') +
  theme(legend.title = element_blank()) +
  theme_classic()

ggsave('figs/fig4.png', plot = last_plot(),
       width = 6, height = 3, units = 'in')

```

<br>

#### Sample-based rarefaction

**Using random method (site-based):**

```{r species accumulation curves}

# Using random method (site-based)

accum_unb <- specaccum(matrix_unb, method = 'random', permutations = 1000,
                       w = effort$effort[1:9])
unb_data <- data.frame(site = accum_unb$sites,
                       effort = accum_unb$effort,
                       richness = accum_unb$richness,
                       SD = accum_unb$sd)

accum_mod <- specaccum(matrix_mod, method = 'random', permutations = 1000,
                       w = effort$effort[9:18])
mod_data <- data.frame(site = accum_mod$sites,
                       effort = accum_mod$effort,
                       richness = accum_mod$richness,
                       SD = accum_mod$sd)

accum_high <- specaccum(matrix_high, method = 'random', permutations = 1000,
                        w = effort$effort[19:27])
high_data <- data.frame(site = accum_high$sites,
                        effort = accum_high$effort,
                        richness = accum_high$richness,
                        SD = accum_high$sd)

# Plot the species accumulation curves:

ggplot() +
  geom_line(data = unb_data,
            aes(x = effort, y = richness),
            size = 1, color = 'darkgreen') +
  geom_ribbon(data = unb_data,
              aes(x = effort, ymin = (richness-2*SD), ymax=(richness+2*SD)),
              alpha = 0.2, fill = 'darkgreen') +
  geom_line(data = mod_data,
            aes(x = effort, y = richness),
            size = 1, color = 'orange') +  
  geom_ribbon(data = mod_data,
              aes(x = effort, ymin = (richness-2*SD), ymax=(richness+2*SD)),
              alpha = 0.2, fill = 'orange') +
  geom_line(data = high_data,
            aes(x = effort, y = richness),
            size = 1, color = 'red') +
  geom_ribbon(data = high_data,
              aes(x = effort, ymin = (richness-2*SD), ymax=(richness+2*SD)),
              alpha = 0.2, fill = 'red') +
  labs(x = 'Trapping Effort',
       y = 'Richness') +
  theme_classic()

```

<br>

#### Per-site richness (individual-based rarefaction estimates)

```{r rarefaction, include=FALSE}

# Calculate richness for each site using rarefaction
raremax <- min(rowSums(matrix))
raremax # minimum sample count = 5: rarefy using this value
rarefied_richness <- rarefy(matrix, sample = raremax)

# Make a dataframe with site, severity, and richness
richness <- tibble(site = matrix_df$site, severity = matrix_df$severity, rarefied_richness)

```

```{r data distribution, eval=FALSE}

# Check summary stats and model assumptions for richness

# Summary table
richness %>% 
  group_by(severity) %>% 
  get_summary_stats(rarefied_richness)

# Check for outliers
richness %>% group_by(severity) %>% 
  identify_outliers(rarefied_richness)
# none

# Check for normality
lm <- lm(rarefied_richness~severity, data=richness)
ggqqplot(residuals(lm))
ggqqplot(richness, "rarefied_richness", facet.by="severity")
shapiro_test(residuals(lm))
richness %>% group_by(severity) %>% 
  shapiro_test(rarefied_richness)
# Looks pretty good

# Homogoneity of variance
plot(lm,1)
richness %>% levene_test(rarefied_richness~severity)
# Variance is fine

# ANOVA
aov <- aov(rarefied_richness~severity, data=richness)
summary(aov)
# p=0.00357
cohens_f(aov)
# effsize=0.77, n=27 (large)
richness %>% ungroup() %>% 
  tukey_hsd(rarefied_richness~severity)
# unb/mod p=0.357, unb/high p=0.00273, mod/high p=0.0332

# Kruskal-Wallis
kruskal_test(richness, rarefied_richness~severity)
# p=0.0131
kruskal_effsize(richness, rarefied_richness~severity)
# effsize=0.278, n=27 (large)
dunn_test(richness, rarefied_richness~severity)
# unb/mod p=0.223, unb/high p=0.0102, mod/high p=0.0872

```

```{r Rarefied Richness boxplot}

# Jitterplot with mean/SE
ggplot(richness, aes(x = severity, y = rarefied_richness)) +
  geom_jitter(aes(color = severity)) +
  scale_color_manual(values = cols) +
  stat_summary(fun.data = mean_se, geom = "crossbar", width = .5) +
  theme_classic() +
  labs(x = 'Fire Severity',
       y = 'Rarefied Richness') +
  theme(legend.position = 'NA') +
  scale_x_discrete(labels = xlabs)

```

- Differences in mean richness calculated from un-pooled rarefaction were significant (one-way ANOVA (2, 24) = 7.1, p < 0.01, n = 9 for each).
- No way to account for sampling effort (uses raw species matrix with counts)

<br>

### Pielou's Evenness (J)

```{r}

# Calculate Pielou's Evenness

J <- diversity(matrix_adj)/log(specnumber(matrix_adj)) # H / log (number of species)

diversity <- cbind(richness, J)
diversity$J[is.nan(diversity$J)] <- NA

```

**Pielou's Evenness (J) decreased across the fire severity gradient:**

```{r J boxplot}

# Jitterplot with mean/SE
ggplot(diversity, aes(x = severity, y = J)) +
  geom_jitter(aes(color = severity)) +
  scale_color_manual(values = cols) +
  stat_summary(fun.data = mean_se, geom = "crossbar", width = .5) +
  theme_classic() +
  labs(x = 'Fire Severity',
       y = 'Pielou\'s Evenness (J)') +
  theme(legend.position = 'NA') +
  scale_x_discrete(labels = xlabs)

```

```{r J data distribution, eval=FALSE}

# Check summary stats and model assumptions for evenness

# Summary table
diversity %>% 
  group_by(severity) %>% 
  get_summary_stats(J)

# Check for outliers
diversity %>% group_by(severity) %>% 
  identify_outliers(J)
# 3 outliers (GRNCAR, GRNKAN, GRNWOO)

# Check for normality
lm <- lm(J~severity, data=diversity)
ggqqplot(residuals(lm))
ggqqplot(diversity, "J", facet.by="severity")
shapiro_test(residuals(lm))
diversity %>% group_by(severity) %>% 
  shapiro_test(J)
# Looks pretty good

# Homogoneity of variance
plot(lm,1)
diversity %>% levene_test(J~severity)
# Variance is fine

#ANOVA
aov <- aov(J~severity, data=diversity)
summary(aov)
# p=0.0445
cohens_f(aov)
# effsize=0.60, n=27 (large)
diversity %>% ungroup() %>% 
  tukey_hsd(J~severity)
# unb/mod p=0.309, unb/high p=0.0375, mod/high p=0.362

# Kruskal-Wallis
kruskal_test(diversity, J~severity)
# p=0.068
kruskal_effsize(diversity, J~severity)
# effsize=0.141, n=27 (large)
dunn_test(diversity, J~severity)
# unb/mod p=0.313, unb/high p=0.0614, mod/high p=0.285

```

<br>

### Shannon Diversity (H)

```{r calculate H}

# Calculate Shannon-Weaver diversity index (H) for each site (adjusted for trapping effort)

H <- diversity(matrix_adj)

diversity <- cbind(diversity, H)

# Export all diversity data

write_csv(diversity, 'output_data/diversity.csv', col_names = TRUE)

```

**Shannon diversity (H) decreased across the fire severity gradient:**

```{r H boxplot}

# Jitterplot with mean/SE
ggplot(diversity, aes(x = severity, y = H)) +
  geom_jitter(aes(color = severity)) +
  scale_color_manual(values = cols) +
  stat_summary(fun.data = mean_se, geom = "crossbar", width = .5) +
  theme_classic() +
  labs(x = 'Fire Severity',
       y = 'Shannon Diversity (H)') +
  theme(legend.position = 'NA') +
  scale_x_discrete(labels = xlabs)

```

```{r H data distribution, eval=FALSE}

# Check summary stats and model assumptions for evenness

# Summary table
diversity %>% 
  group_by(severity) %>% 
  get_summary_stats(H)

# Check for outliers
diversity %>% group_by(severity) %>% 
  identify_outliers(H)

# Check for normality
lm <- lm(H~severity, data=diversity)
ggqqplot(residuals(lm))
ggqqplot(diversity, "H", facet.by="severity")
shapiro_test(residuals(lm))
diversity %>% group_by(severity) %>% 
  shapiro_test(H)
# Looks pretty good

# Homogoneity of variance
plot(lm,1)
diversity %>% levene_test(H~severity)
# Variance is fine

#ANOVA
aov <- aov(H~severity, data=diversity)
summary(aov)
diversity %>% ungroup() %>% 
  tukey_hsd(H~severity)
cohens_f(aov)

```

- Diversity differed significantly between the three fire severity treatments (one-way ANOVA (2, 24) = 6.2, p < 0.01, n = 9 for each). Post-hoc analysis by Tukey’s HSD showed that mean diversity in unburned sites (H = 0.87) and high severity sites (H = 0.32) was significantly different (pairwise p = 0.005). 

<br>

```{r eval=FALSE}
rmarkdown::render('code/04_diversity_and_richness.Rmd', output_file = '../docs/04_diversity_and_richness.html')
```
