==> code/05_abundance.Rmd <==
---
title: "Abundance Analyses"
subtitle: "From [SMAMMUNITY: SMall mAMMal commUNITY](https://github.com/katekathrynkat/smammunity)"
output:
  html_document:
    toc: true
    toc_float: true
    theme: flatly
    highlight: tango
    code_folding: hide
editor_options: 
  chunk_output_type: console
---

```{r setup, include=FALSE}
knitr::opts_knit$set(root.dir = '..')
knitr::opts_chunk$set(echo = FALSE, warning = FALSE, message = FALSE,
                      fig.width=5, fig.height=3)
```

**Kate Culhane**, Dept. of Ecology, Evolution, and Marine Biology, UC Santa Barbara\
[kathrynculhane\@ucsb.edu](mailto:kathrynculhane@ucsb.edu){.email}

**Publication:** Culhane, K., R. Sollmann, A.M. White, G.L. Tarbill, S. Cooper, M.A. Moritz, H.S. Young. (2021). Small mammal functional traits and vegetation characteristics mediate community responses to fire severity. Ecological Applications. Manuscript submitted for initial review.

**Source code:** [05_abundance.Rmd](https://github.com/katekathrynkat/smammunity/blob/master/code/05_abundance.Rmd)

---

```{r packages and data}

# Packages
library(rstatix) # KW test, KW effect size
library(ggpubr) # 
library(dunn.test) # Dunn's test (KW post hoc)
library(effectsize) # 
library(kableExtra) # making pretty tables
library(calecopal) # color palettes
library(tidyverse)

# Small mammal unique individuals
smamms <- read_csv('output_data/02_unique_smamms.csv') %>% 
  mutate(severity = factor(severity, levels = c('unb', 'mod', 'high')))

# Small mammal capture data
captures <- read_csv('output_data/02_smamms.csv') %>% 
  mutate(severity = factor(severity, levels = c('unb', 'mod', 'high')))

# Small mammal trapping effort
effort <- read_csv('output_data/02_effort.csv') %>% 
  mutate(severity = factor(severity, levels = c('unb', 'mod', 'high')))

# Site metadata
site_meta <- read_csv('raw_data/field_data/site_metadata.csv') %>% 
  mutate(severity = factor(severity, levels = c('unb', 'mod', 'high')))

# Vectors for plotting
cols <- c('#4AB793', '#DCA827', '#D85F2B')
xlabs <- c('Unburned', 'Moderate Severity', 'High Severity') 

```

```{r}

# Calculate total abundance summed across all species at each site (raw and adjusted)

abundance <- smamms %>% 
  group_by(site, severity) %>% 
  tally() %>% # count individuals per site
  full_join(effort) %>% 
  mutate(n_adj = n/effort*300) %>% # add column adjusted for effort
  ungroup() %>% 
  full_join(site_meta[2:4]) %>% 
  mutate(offset=effort/300)

```

## Total unique individuals per species

```{r}

# Summary table of overall mammal abundances

summary <- smamms %>% 
  group_by(species, binomial, family) %>% 
  summarize(n = length(species)) %>% 
  ungroup() %>% 
  arrange(-n)

kable(summary[,2:4], col.names = c('Binomial', 'Family', 'Individuals')) %>% 
  kable_styling(bootstrap_options = c('striped', 'condensed'),
                full_width = FALSE,
                position = 'left') %>% 
  column_spec(1, italic = TRUE) %>% 
  column_spec(1, width = '5cm')

```

- Although traps were primarily open at night, we can still include diurnal species (chipmunks, squirrels) in measures of relative abundance because traps were opened and closed at approximately the same time each day.

- Flying squirrels spend a significant amount of time foraging on the ground, so their capture was not a fluke.

---

## Ratio of recaptured animals

```{r}

# Ratio of recaps: caps
# Comparison between sites

ratio <- captures %>% 
  filter(species!='SPRU' & species!='MISS') %>% 
  group_by(site) %>% 
  tally() %>% 
  full_join(abundance[1:3], by=c('site')) %>% 
  mutate(rate=n.y/n.x)

ggplot(ratio, aes(x = severity, y = rate)) +
  geom_jitter(aes(color = severity)) +
  stat_summary(fun.data = mean_se, geom = "crossbar", width = .5) +
  theme_classic() +
  labs(x = 'Fire Severity',
       y = 'Recap Rate') +
  theme(legend.position = 'NA')

kruskal.test(rate~severity, data=ratio)

```

- Ratio of recaptures : total captures for each site

- Very similar between treatments (sort of all over the place, though)

---

## Proportional abundance

### By site

```{r proportional abundance per site, fig.width=7}

# Dataframe of proportional species abundance per site
prop_site <- smamms %>% 
  group_by(site, severity, species, binomial) %>% 
  summarize(n = length(species)) %>% 
  group_by(site) %>% 
  mutate(prop = n/sum(n)*100) %>% 
  ungroup() %>% 
  arrange(severity) %>% 
  mutate(site = factor(site, levels = unique(site))) %>% # order sites by severity
  mutate(binomial = factor(binomial, levels = summary$binomial)) # order species by total abundance

# Column graph of proportional species abundance per site
ggplot(prop_site, aes(x = site, y = prop,
                                       fill = binomial)) +
  geom_col() +
  labs(x = 'Site',
       y = 'Occurence by species (%)') +
  theme(axis.text.x = element_text(angle = 60, hjust = 1)) +
  scale_y_continuous(breaks = c(0, 50, 100),
                     expand = c(0, 0)) +
  scale_fill_manual(values=cal_palette(name = 'sierra2', n = 11, type = 'continuous')) +
  theme_minimal() +
  theme(panel.grid.major = element_blank(),
        panel.grid.minor = element_blank(),
        axis.text.x = element_text(angle = 90, hjust = 1)) +
  facet_grid(~ severity,
             scale = 'free',
             labeller = as_labeller(c('unb' = 'Unburned',
                                      'mod' = 'Moderate',
                                      'high' = 'High')))

```

<br>

### By species

```{r proportional abundance per species}

# Data frame of proportional treatment abundances per species

prop_species <- smamms %>% 
  group_by(species, common_name, site, severity) %>% 
  summarize(n = length(site)) %>% 
  full_join(effort) %>% 
  mutate(cpue = n/effort * 300) %>% # adjust for trapping effort
  group_by(species, common_name, severity) %>% 
  summarize(n = sum(cpue)) %>% 
  mutate(prop = n/sum(n)*100) %>% # proportion of captures at each treatment, per species
  mutate(label = common_name) %>% 
  ungroup()

prop_low <- prop_species %>%  # data frame for just low severity
  filter(severity == 'unb') %>% 
  select(species, prop) %>% 
  rename(prop_low = prop)
prop_mod <- prop_species %>%  # data for just moderate severity
  filter(severity == 'mod') %>% 
  select(species, prop) %>% 
  rename(prop_mod = prop)

prop_species2 <- prop_species %>% # re-ordering data frame occurence within each severity
  full_join(prop_low) %>% 
  full_join(prop_mod) %>% 
  mutate(
    prop_low = case_when(
      is.na(prop_low) ~ 0,
      !is.na(prop_low) ~ prop_low
    ),
    prop_mod = case_when(
      is.na(prop_mod) ~ 0,
      !is.na(prop_mod) ~ prop_mod
    )
  ) %>% 
  ungroup() %>% 
  mutate(label = as_factor(label),
         label = fct_reorder(label, prop_mod),
         label = fct_reorder(label, prop_low),
         severity = factor(severity, levels = c('high', 'mod', 'unb'))) %>% 
  full_join(summary, by = 'species')

# Column graph of proportional treatment abundances per species

colors <- c("#D85F2B", "#DCA827", "#2D9875")

ggplot(prop_species2, aes(x = label, y = prop, fill = severity)) +
  geom_col() +
  geom_text(aes(x = label, y = 103,
                label = paste('n = ', n.y)),
            size = 3.2, hjust = 0, color = 'grey30') +
  scale_fill_manual(values = colors,
                    name = 'Fire severity',
                    labels = labs) +
  coord_flip() +
  scale_y_continuous(limits = c(0, 200),
                     breaks = c(0, 50, 100),
                     expand = c(0,0)) +
  theme_minimal() +
  labs(x = '', y = '') +
  theme(axis.text.y = element_text(size = 10, color = 'black'),
        axis.text.x = element_text(size = 10),
        panel.grid.major = element_blank(),
        panel.grid.minor = element_blank(),
        legend.position = 'none',
        axis.title.x = element_text(hjust = 3.8, size = 13))

ggsave('figs/05_fig4a.png', plot = last_plot(),
       width = 4, height = 3, units = 'in')

```

- Habitat preferences for the 11 species of mammals, standardized by trapping effort.

---

## Total abundance

### Abundance

```{r}

# Jitterplot + boxplot
ggplot(abundance, aes(x = severity, y = n_adj)) +
  geom_jitter(aes(color = severity, shape = severity), width = 0.2, size = 2) +
  scale_color_manual(values = cols) +
  scale_shape_manual(values=c(19,17,15)) +
  geom_boxplot(outlier.shape=NA, fill=NA) +
  theme_classic() +
  labs(x = 'Fire Severity',
       y = 'Total small mammal abundance') +
  theme(legend.position = 'NA') +
  scale_x_discrete(labels = xlabs)

```

```{r, eval=FALSE}

# Check summary stats and model assumptions for total abundance per site

# Summary table
abundance %>% 
  group_by(severity) %>% 
  get_summary_stats(n_adj)

# Check for outliers
abundance %>% group_by(severity) %>% 
  identify_outliers(n_adj)
# 2 outliers (HICORN and MIXMOR), no extreme outliers

# Check for normality
lm <- lm(n_adj~severity, data=abundance)
ggqqplot(residuals(lm))
ggqqplot(abundance, "n_adj", facet.by="severity")
shapiro_test(residuals(lm))
abundance %>% group_by(severity) %>% 
  shapiro_test(n_adj)
# The two outliers make the data not normal

# Homogoneity of variance
plot(lm,1)
abundance %>% levene_test(n_adj~severity)
# Variance is fine

#Kruskal-Wallis
kruskal_test(n_adj~severity, data=abundance)
# p=0.0658
kruskal_effsize(n_adj~severity, data=abundance)
# effsize=0.143, n=27 (large)

```

```{r}

kruskal.test(n_adj~severity, data=abundance)
kruskal_effsize(n_adj~severity, data=abundance)

```

- Jitter plot points represent total small mammal abundance at each site; error bars represent mean and standard error for each fire severity category.

- Abundance was standardized by trapping effort.

- Differences marginally significant by K-W (p=0.06, large eff size)

<br>

### Biomass

```{r}

# Average mass for each unique individual
capture_mass <- captures %>% 
  select(site, severity, indID, binomial, weight, age) %>% 
  filter(!is.na(indID), # filter for captures only
         (age == 'A' | binomial == 'Neotamias amoenus')) %>% # filter for only adult animals (all NEAM were subadults, unfortunately)
  group_by(indID, site, severity, binomial) %>% 
  summarize(avg_mass = mean(weight, na.rm = TRUE)) %>% 
  ungroup()

# Average mass for each species
species_mass <- capture_mass %>% 
  group_by(binomial) %>% 
  summarize(spp_avg = mean(avg_mass, na.rm = TRUE))

# Average mass for each unique individual, with NAs filled in with species averages
capture_biomass <- full_join(capture_mass, species_mass, by = 'binomial') %>% 
  mutate(mass = case_when(
    !is.na(avg_mass) ~ avg_mass,
    is.na(avg_mass) ~ spp_avg
  )) %>% 
  select(site, severity, binomial, mass)

# Total biomass at each site
biomass <- capture_biomass %>% 
  group_by(site, severity) %>% 
  summarize(biomass = sum(mass)) %>% 
  ungroup()

```

```{r}

# Jitterplot with mean/SE
ggplot(biomass, aes(x = severity, y = biomass)) +
  geom_jitter(aes(color = severity, shape = severity), width = 0.2, size = 2) +
  scale_color_manual(values = cols) +
  scale_shape_manual(values=c(19,17,15)) +
  stat_summary(fun.data = mean_se, geom = 'errorbar', width = 0.2, size = 0.7) +
  stat_summary(fun = 'mean', size = 2, geom = 'point') +
  theme_classic() +
  labs(x = 'Fire Severity',
       y = 'Total small mammal biomass') +
  theme(legend.position = 'NA') +
  scale_x_discrete(labels = xlabs)

```

```{r, eval=FALSE}

# Check summary stats and model assumptions for total biomass per site

# Summary table
biomass %>% 
  group_by(severity) %>% 
  get_summary_stats(biomass)

# Check for outliers
biomass %>% group_by(severity) %>% 
  identify_outliers(biomass)
# 2 outliers (MIXMOR, HICAMP)

# Check for normality
lm <- lm(biomass~severity, data=biomass)
ggqqplot(residuals(lm))
ggqqplot(biomass, "biomass", facet.by="severity")
shapiro_test(residuals(lm))
# The outliers make the data not normal

# Homogoneity of variance
plot(lm,1)
biomass %>% levene_test(biomass~severity)
# Variance is fine

#Kruskal-Wallis
kruskal.test(biomass~severity, data=biomass)
# p=0.3446
kruskal_effsize(biomass~severity, data=biomass)
# effsize=0.00544, n=27 (small)

```

```{r}

kruskal.test(biomass~severity, data=biomass)
kruskal_effsize(biomass~severity, data=biomass)

```

- Similar pattern, but low/mod sites seem to have proportionally high biomass (probably because of ground squirrels)

- NOT significant by K-W (p = 0.48, small eff size)

---

## Abundance by species

```{r abundance per species, include=TRUE, fig.width=2.5, fig.height=2}

# Plot total abundance for species at each site

for (i in unique(smamms$species)) {
  
  data <- smamms %>% 
  filter(species == i) %>% 
  group_by(site, severity) %>% 
  tally() %>%
  full_join(abundance, by = c('site', 'severity')) %>% 
  mutate(non_target = (n.y-n.x),
         effort_target = effort - non_target/2,
         n_target = n.x,
         n_adj_target = case_when(
           !is.na(n_target) ~ (n_target/effort_target)*300,
           TRUE ~ 0
           )) %>% 
  select(site, severity, n_target, effort_target, n_adj_target) 
  
  plot <- ggplot(data, aes(x = severity, y = n_adj_target)) +
    geom_jitter(aes(color = severity), height=0) +
    scale_color_manual(values = cols) +
    stat_summary(fun.data = mean_se, geom = "crossbar", width=0.5) +
    theme_classic() +
    labs(x = 'Fire Severity',
         y = 'Abundance',
         title = i) +
    theme(legend.position = 'NA') +
    scale_x_discrete(labels = xlabs)
  
  print(plot)
  
}

```

- Jitter plot points represent total small mammal abundance at each site; horizontal bars represent mean and standard error for each fire severity category.

- Abundance was standardized by trapping effort for each species individually (i.e. non-target species were counted as half of a trap night, similar to treatment of sprung traps).

- Two of these sites look like outliers:
  - HICORN had a ridiculous number of PEMA
  - MIXMOR had a lot of PEMA, OTBE, and SOTR

<br>

### Deer mouse (PEMA) abundance
  
```{r}

# PEMA abundance
pema <- smamms %>% 
  filter(species == 'PEMA') %>% 
  group_by(site, severity) %>% 
  tally() %>%
  full_join(abundance, by = c('site', 'severity')) %>% 
  mutate(non_target = (n.y-n.x),
         effort_pema = effort - non_target/2,
         n_pema = n.x,
         n_adj_pema = case_when(
           !is.na(n_pema) ~ (n_pema/effort_pema)*300,
           TRUE ~ 0
           ),
         offset=effort_pema/300) %>% 
  select(site, severity, block, n_pema, offset, n_adj_pema) %>%
  ungroup()

# Jitterplot with mean/SE
ggplot(pema, aes(x = severity, y = n_adj_pema)) +
  geom_jitter(aes(color = severity)) +
  scale_color_manual(values = cols) +
  stat_summary(fun.data = mean_se, geom = "crossbar", width = .5) +
  theme_classic() +
  labs(x = 'Fire Severity',
       y = 'PEMA Abundance') +
  theme(legend.position = 'NA') +
  scale_x_discrete(labels = xlabs)

```

```{r eval=FALSE}

# Check summary stats and model assumptions for total PEMA abundance per site

# Summary table
pema %>% 
  group_by(severity) %>% 
  get_summary_stats(n_adj_pema)

# Check for outliers
pema %>% group_by(severity) %>% 
  identify_outliers(n_adj_pema)
# 1 outlier in HICORN, no extreme outliers

# Check for normality
lm <- lm(n_adj_pema~severity, data=pema)
ggqqplot(residuals(lm))
ggqqplot(pema, "n_adj_pema", facet.by="severity")
shapiro_test(residuals(lm))
pema %>% 
  group_by(severity) %>% 
  shapiro_test(n_adj_pema)
# The outlier makes the data not normal

# Homogoneity of variance
plot(lm,1)
pema %>% ungroup() %>%
  levene_test(n_adj_pema~severity)
# Variance is fine

# Kruskal-Wallis test
kruskal.test(n_adj_pema~severity, data=pema)
# p=0.00982
kruskal_effsize(n_adj_pema~severity, data=pema)
# effsize=0.302, p=27 (large)
dunn.test(pema$n_adj_pema, pema$severity, method='bonferroni')
# unb/mod p=0.859, unb/high p=0.0197, mod/high p=0.0222

```

```{r}

# Stats
dunn.test(pema$n_adj_pema, pema$severity, method='bonferroni')
kruskal_effsize(n_adj_pema~severity, data=pema)

```

- Jitter plot points represent PEMA abundance at each site; horizontal bars represent mean and standard error for each fire severity category. Abundance was standardized by trapping effort.

- Significant by K-W (p = 0.01, large eff size)

<br>

### Trowbridge's shrew (SOTR) abundance

```{r}

# SOTR abundance
sotr <- smamms %>% 
  filter(species == 'SOTR') %>% 
  group_by(site, severity) %>% 
  tally() %>%
  full_join(abundance, by = c('site', 'severity')) %>% 
  mutate(non_target = (n.y-n.x),
         effort_sotr = effort - non_target/2,
         n_sotr = n.x,
         n_adj_sotr = case_when(
           !is.na(n_sotr) ~ (n_sotr/effort_sotr)*300,
           TRUE ~ 0
           ),
           offset=effort_sotr/300) %>% 
  select(site, severity, block, n_sotr, offset, n_adj_sotr) %>%
  ungroup()

# Jitterplot with mean/SE
ggplot(sotr, aes(x = severity, y = n_adj_sotr)) +
  geom_jitter(aes(color = severity)) +
  scale_color_manual(values = cols) +
  stat_summary(fun.data = mean_se, geom = "crossbar", width = .5) +
  theme_classic() +
  labs(x = 'Fire Severity',
       y = 'SOTR Abundance') +
  theme(legend.position = 'NA') +
  scale_x_discrete(labels = xlabs)

```

```{r eval=FALSE}

# Check summary stats and model assumptions for total PEMA abundance per site

# Summary table
sotr %>% 
  group_by(severity) %>% 
  get_summary_stats(n_adj_sotr)

# Check for outliers
sotr %>% group_by(severity) %>% 
  identify_outliers(n_adj_sotr)
# 2 outliers: MIXDP2 and MIXMOR, both extreme

# Check for normality
lm <- lm(n_adj_sotr~severity, data=sotr)
ggqqplot(residuals(lm))
ggqqplot(sotr, "n_adj_sotr", facet.by="severity")
shapiro_test(residuals(lm))
# The outlier makes the data not normal

# Homogoneity of variance
plot(lm,1)
sotr %>% ungroup() %>%
  levene_test(n_adj_sotr~severity)
# Variance is all weird because high sites had no shrews

# Kruskal-Wallis test
kruskal.test(n_adj_sotr~severity, data=sotr)
# p=0.00009407
kruskal_effsize(n_adj_sotr~severity, data=sotr)
# effsize=0.686, p=27 (large)
dunn.test(sotr$n_adj_sotr, sotr$severity, method='bonferroni')
# unb/mod p=0.0006, unb/high p=0.0000, mod/high p=0.1929

```

```{r}

# Stats

dunn.test(sotr$n_adj_sotr, sotr$severity, method='bonferroni')
kruskal_effsize(n_adj_sotr~severity, data=sotr)

```

- Jitter plot points represent SOTR abundance at each site; horizontal bars represent mean and standard error for each fire severity category. Abundance was standardized by trapping effort.

- Significant by K-W (p < 0.001, large eff size)

```{r include=FALSE}

# Export abundance data

abundance_dat <- full_join(abundance, pema, by = c('site', 'severity')) %>% 
  full_join(sotr, by = c('site', 'severity')) %>% 
  full_join(biomass, by = c('site', 'severity')) %>% 
  select(site, severity, n, n_adj,
         n_pema, n_adj_pema,
         n_sotr, n_adj_sotr,
         biomass)

write_csv(abundance_dat, 'output_data/05_abundance.csv', col_names = TRUE)

```

```{r eval=FALSE}
rmarkdown::render('code/05_abundance.Rmd', output_file = '../docs/05_abundance.html')
```
