==> code/08_nmds.Rmd <==
---
title: "NMDS"
author: "Kate Culhane"
output: html_document
editor_options: 
  chunk_output_type: console
---

```{r setup, include=FALSE}
knitr::opts_knit$set(root.dir = '..')
knitr::opts_chunk$set(echo = FALSE, warning = FALSE, message = FALSE,
                      fig.width=7, fig.height=5)
```

<br>

```{r packages and data}

# Load packages

library(tidyverse)
library(vegan)
library(goeveg)
library(pairwiseAdonis)
library(ggvegan)
library(ggrepel)

# Update functions

select <- dplyr::select

# Load data

matrix_df <- read_csv('output_data/species_matrix.csv') %>% 
  mutate(severity = factor(severity, levels = c('unb', 'mod', 'high')))

matrix_adj_df <- read_csv('output_data/species_matrix_adj.csv') %>% 
  mutate(severity = factor(severity, levels = c('unb', 'mod', 'high')))

site_meta <- read_csv('raw_data/field_data/site_metadata.csv') %>% 
  mutate(severity = factor(severity, levels = c('unb', 'mod', 'high')))

veg_metrics <- read_csv('output_data/veg_metrics.csv') %>% 
  mutate(severity = factor(severity, levels = c('unb', 'mod', 'high')))

smamm_meta <- read_csv('raw_data/lit_data/mammal_traits.csv')

```

```{r data wrangling}

# Transform data frames into matrices

matrix <- as.matrix(matrix_df[-c(1,2)])
rownames(matrix) <- matrix_df$site

matrix_adj <- as.matrix(matrix_adj_df[-c(1,2)])
rownames(matrix_adj) <- matrix_adj_df$site

matrix_hell <- decostand(matrix, 'hellinger')

matrix_sqrt <- sqrt(matrix)

# Severity data frame

sevs <- select(matrix_df, site, severity)

```

### Community structure

For choosing transformation and standardization methods: https://chrischizinski.github.io/SNR_R_Group/2016-08-10-Data-Transformations


```{r nmds}

# NMDS with relative abundance data (2 dimensions)
set.seed(17)
ord <- metaMDS(matrix_adj,
               distance = 'bray',
               autotransform = FALSE,
               k = 2,
               try=50)
ord$stress # 0.09
stressplot(ord)

# Scree plot to check stress per number of dimensions
dimcheckMDS(matrix_hell, distance = "bray", autotransform = FALSE, k = 10, trymax = 20)

```

**NMDS plot showing community structure across the three fire severities:** 

```{r nmds plot, echo=FALSE}

cols <- c('darkgreen', 'orange', 'red')
par(mar = c(4,4,1,1))
plot(ord, display = 'species', type = 'n')
points(ord, display = 'sites',
       cex = 1.5, pch = 19, col = cols[sevs$severity])
ordiellipse(ord, groups = sevs$severity,
            label = TRUE, col = cols, lwd = 2)
text(ord, display = 'species',
     cex = 0.6)

```

```{r ANOSIM and adonis, eval=FALSE}

# Attach severity data to species matrix
matrix_hell
attach(sevs)

# Compare community similarity across fire severities using ANOSIM
smamm_ano <- anosim(matrix_hell, severity)
summary(smamm_ano)
plot(smamm_ano)

# Compare community dispersion across fire severities using betadisper
smamm_dist <- vegdist(matrix_hell)
smamm_bd <- betadisper(smamm_dist, severity)

smamm_bd
plot(smamm_bd)
boxplot(smamm_bd)

permutest(smamm_bd)
anova(smamm_bd)
# p > 0.05 for both --> no difference in dispersion across fire severities
# Satisfies the assumption for adonis


# Compare community similarity across fire severities using adonis
# Adonis is more robust because it is less sensitive to dispersion
smamm_ado <- adonis(matrix_hell ~ severity,
                    permuatations = 999,
                    method = 'bray')
smamm_ado

# Pairwise Adonis
smamm_ado_pw <- pairwise.adonis(matrix_hell, severity,
                                sim.method = 'bray',
                                p.adjust.m = 'holm')

smamm_ado_pw

```


        pairs Df  SumsOfSqs    F.Model         R2 p.value p.adjusted sig
    1  unb vs mod  1 0.31713367  4.9151490 0.23500425   0.002      0.006   *
    2 unb vs high  1 0.45329584 10.2675595 0.39088365   0.001      0.003   *
    3 mod vs high  1 0.05364198  0.9347399 0.05519659   0.466      1.000    

- NMDS in two dimensions using relative abundance species matrix and Wisconsin correction on Bray-Curtis distances
- Appears to be fewer sites because some of them exactly overlap
- Adonis F=4.739, p=0.002
- Pairwise Adonis unb vs. mod p=0.018, unb vs high p=0.003 (Bonferroni correction)
- Considered using ANOSIM instead of Adonis, but Adonis is more robust because it is less sensitive to dispersion (also used betadisper to determine that dispersion was similar across treatments in order to satisfy assumptions of Adonis)
- Species codes:
    - SPBE = California ground squirrel
    - GLSA = northern flying squirrel
    - TAAM = yellow-pine chipmunk
    - TAQU = long-eared chipmunk
    - TASE = shadow chipmunk
    - NEFU = dusky-footed woodrat
    - PEMA = North American deer mouse
    - PEBO = brush mouse
    - PETR = Pinyon mouse
    - REME = western harvest mouse
    - SOTR = Trowbridge's shrew

<br>

### Habitat drivers

```{r model nmds}

# Data frame with only the variables from the model
model_vars <- veg_metrics %>% 
  select(site, severity,
         litter_cover, live_tree_density, soft_cwd_volume,
         shrub_cover, forb_grass_cover, pca_score)

# Fit environmental variables
fit <- envfit(ord, model_vars[,3:8], permutations = 1000)

# NMDS plot with arrows for environmental variables
par(mar = c(4,4,1,1))
plot(ord, diplay = 'sites', type = 'n')
points(ord, display = 'sites',
       cex = 1, pch = 19, col = cols[sevs$severity])
ordiellipse(ord, groups = sevs$severity, # ellipse with std error around centroids
            label = FALSE, col = cols, lwd = 2)
plot(fit,
     col = 'black', lwd = 3, cex = 0.8)
text(ord, display = 'species',
     cex = 0.7, col = 'grey30')

```

<br> 

**Clean NMDS plots with ggplot:**

#### Data

```{r, include=FALSE}

# Ordination plot with ggplot

# Data frames with point and vector informations

site_points <- fortify(ord) %>% 
  filter(Score == 'sites') %>% 
  full_join(sevs, by = c('Label' = 'site')) %>% 
  dplyr::rename(group = severity)


spp_points <- fortify(ord) %>% 
  filter(Score == 'species')

vectors <- fortify(fit) %>% 
  mutate(xend=NMDS1*1.5,
         yend=NMDS2*1.5,
         xlab = case_when(
           Label == 'shrub_cover' ~ xend+0.33,
           Label == 'litter_cover' ~ xend+0.2,
           Label == 'live_tree_density' ~ xend-0.34,
           Label == 'soft_cwd_volume' ~ xend-0.2,
           Label == 'forb_grass_cover' ~ xend+0.49,
           Label == 'pca_score' ~ xend+0.3
         ),
         ylab = case_when(
           Label == 'shrub_cover' ~ yend+0.04,
           Label == 'litter_cover' ~ yend-0.06,
           Label == 'live_tree_density' ~ yend,
           Label == 'soft_cwd_volume' ~ yend+0.08,
           Label == 'forb_grass_cover' ~ yend-0.03,
           Label == 'pca_score' ~ yend
         ),
         lab = case_when(
           Label == 'shrub_cover' ~ 'Shrub cover',
           Label == 'litter_cover' ~ 'Litter cover',
           Label == 'live_tree_density' ~ 'Tree density',
           Label == 'soft_cwd_volume' ~ 'Soft CWD',
           Label == 'forb_grass_cover' ~ 'Forb+grass cover',
           Label == 'pca_score' ~ 'PC1[T+S+L]'
         ))

# Data frame with ellipse information

plot.new()
ord_el <- ordiellipse(ord, groups = sevs$severity)

veganCovEllipse<-function (cov, center = c(0, 0), scale = 1, npoints = 100) 
  {
    theta <- (0:npoints) * 2 * pi/npoints
    Circle <- cbind(cos(theta), sin(theta))
    t(center + scale * t(Circle %*% chol(cov)))
  }

df_ell <- data.frame()

for(g in levels(site_points$group)){
  df_ell <- rbind(df_ell, cbind(as.data.frame(with(site_points[site_points$group==g,],
                  veganCovEllipse(ord_el[[g]]$cov,ord_el[[g]]$center,ord_el[[g]]$scale)))
                                ,group=g))
}

```

#### Plot

```{r}
  
# Plot
ggplot() +
  geom_path(data=df_ell, aes(x=NMDS1, y=NMDS2, color=group),
            size=0.6) +
  geom_point(data=site_points, aes(x=NMDS1, y=NMDS2, color=group, shape=group),
  size=1.8) +
  geom_segment(data=vectors[1:5,], aes(x=0, y=0, xend=xend, yend=yend),
               arrow=arrow(length=unit(0.02, 'npc'), type='closed'), size=0.6) +
  geom_segment(data=vectors[6,], aes(x=0, y=0, xend=xend, yend=yend),
               arrow=arrow(length=unit(0.02, 'npc'), type='closed'), size=0.6, linetype=5) +
  geom_text(data=vectors[1:5,], aes(x=xlab, y=ylab, label=lab), size=4) +
  geom_text(data=vectors[6,], aes(x=xlab, y=ylab, label=lab), size=4, parse=TRUE) +
  geom_label(data=site_points, aes(x=0.82, y=-0.7,
                                   label='NMDS in 3 dim\n2 dim displayed\nstress = 0.09'),
             size=4) +
  scale_color_manual(values=c('#4AB793', '#DCA827', '#D85F2B')) +
  scale_shape_manual(values=c(19,17,15)) +
  scale_x_continuous(limits=c(-1.6,1.15)) +
  scale_y_continuous(limits=c(-0.9,0.71)) +
  theme_classic() +
  theme(panel.border=element_rect(fill=NA),
        axis.text.x=element_blank(), axis.text.y=element_blank(),
        legend.position='none') +
  coord_fixed(clip='off')

ggsave('figs/fig5b.png', plot=last_plot(),
       width=4.5, height=3.5, units='in')

```

### Traits

#### Data

```{r model nmds}

# Data frame with mammal traits
traits <- read_csv('raw_data/lit_data/mammal_traits_NEW.csv') %>% 
  arrange(species)

# Fit traits
trait_fit <- envfit(ord, traits[,c(4:6)], permutations = 1000, display = 'sp')

# NMDS plot with arrows for traits
par(mar = c(4,4,1,1))
plot(ord, diplay = 'sites', type = 'n')
points(ord, display = 'sites',
       cex = 1, pch = 19, col = cols[sevs$severity])
ordiellipse(ord, groups = sevs$severity, # ellipse with std error around centroids
            label = FALSE, col = cols, lwd = 2)
plot(trait_fit,
     col = 'black', lwd = 3, cex = 0.8)
text(ord, display = 'species',
     cex = 0.7, col = 'grey30')

vectors2 <- fortify(trait_fit) %>% 
  mutate(xend=NMDS1,
         yend=NMDS2,
         lab = case_when(
           Label == 'nestsBurrow' ~ 'Burrow',
           Label == 'nestsHollow' ~ 'Hollow',
           Label == 'nestsTree' ~ 'Tree',
           Label == 'guildInsectivore' ~ 'Insectivore',
           Label == 'guildOmnivore'~ 'Omnivore',
           Label == 'guildHerbivore' ~ 'Herbivore',
           Label == 'foragArboreal' ~ 'Arboreal',
           Label == 'foragGround' ~ 'Ground',
           Label == 'foragScansorial' ~ 'Scansorial'
         ))

vectors3 <- vectors %>% 
  mutate(
    xlab = case_when(
      Label=='soft_cwd_volume'~xlab+0.22,
      Label=='pca_score'~xlab,
      TRUE~xlab
    ),
    ylab = case_when(
      Label=='soft_cwd_volume'~ylab,
      Label=='pca_score'~ylab,
      TRUE~ylab
    )
  )

```

<br>

**Clean NMDS plots with ggplot:**

#### Plot

```{r}

# Plot

ggplot() +
  geom_path(data=df_ell, aes(x=NMDS1, y=NMDS2, color=group),
            size=0.6) +
  geom_segment(data=vectors2, aes(x=0, y=0, xend=NMDS1, yend=NMDS2),
               arrow=arrow(length=unit(0.02, 'npc'), type='closed'), size=0.6) +
  geom_text(data=vectors2, aes(x=NMDS1, y=NMDS2, label=lab),
            size=4) +
  geom_segment(data=vectors3[c(3,6),], aes(x=0, y=0, xend=xend, yend=yend,),
               arrow=arrow(length=unit(0.02, 'npc'), type='closed'),
               size=0.6, linetype=5) +
  geom_text(data=vectors3[c(3,6),], aes(x=xlab, y=ylab, label=lab),
            size=4, fontface='bold') +
  geom_label(data=site_points, aes(x=0.82, y=-0.7,
                                   label='NMDS in 3 dim\n2 dim displayed\nstress = 0.09'),
             size=4) +
  scale_color_manual(values=c('#4AB793', '#DCA827', '#D85F2B')) +
  scale_shape_manual(values=c(19,17,15)) +
  scale_x_continuous(limits=c(-1.6,1.15)) +
  scale_y_continuous(limits=c(-0.9,0.71)) +
  theme_classic() +
  theme(panel.border=element_rect(fill=NA),
        axis.text.x=element_blank(), axis.text.y=element_blank(),
        legend.position='none') +
  coord_fixed(clip='off')

ggsave('figs/fig5b2.png', plot=last_plot(),
       width=4.5, height=3.5, units='in')

```

```{r}

# Species labels

spp_points2 <- spp_points %>% 
  full_join(smamm_meta[c(1,5)], by = c('Label'='species')) %>% 
  mutate(xdat=NMDS1*0.61,
         ydat=NMDS2*0.7)

ggplot() +
  geom_text_repel(data=spp_points2, aes(x=-1, y=ydat, label=common_name),
                  direction='y', segment.alpha=0, hjust=0,
                  size=3.5, color='grey40', point.padding = 0) +
  scale_x_continuous(limits=c(-1.6,1.15)) +
  scale_y_continuous(limits=c(-0.9,0.7)) +
  theme_classic() +
  theme(panel.border=element_rect(fill=NA),
        axis.text.x=element_blank(), axis.text.y=element_blank(),
        legend.position='none') +
  coord_fixed(clip='off')


ggsave('figs/nmdsspecies1.png', plot=last_plot(),
       width=4.5, height=3.5, units='in')

ggplot() +
  geom_text_repel(data=spp_points2, aes(x=xdat, y=-0.7, label=common_name),
                  direction='x', segment.alpha=0, vjust=0,
                  size=3.5, color='grey40', angle=90, point.padding = 0) +
  scale_x_continuous(limits=c(-1.6,1.15)) +
  scale_y_continuous(limits=c(-0.9,0.7)) +
  theme_classic() +
  theme(panel.border=element_rect(fill=NA),
        axis.text.x=element_blank(), axis.text.y=element_blank(),
        legend.position='none') +
  coord_fixed(clip='off')

ggsave('figs/nmdsspecies2.png', plot=last_plot(),
       width=4.5, height=3.5, units='in')
  
```

```{r eval=FALSE}
rmarkdown::render('code/08_nmds.Rmd', output_file = '../docs/08_nmds.html')
```
