==> code/08_vegetation_metrics.Rmd <==
---
title: "Vegetation Metrics"
author: "Kate Culhane"
output:
  html_document:
    toc: true
    toc_float: true
editor_options: 
  chunk_output_type: console
---

```{r setup, include=FALSE}
knitr::opts_knit$set(root.dir = '..')
knitr::opts_chunk$set(echo=FALSE, warning=FALSE, message=FALSE)
```

<br>

*Vegetation metrics hypothesized to a) shift after fire according to fire severity and b) influence small mammal community structure. Variables incorporated into GLM~veg~.*

<br>

### Metrics

```{r include=FALSE}

# Packages
library(vegan) # community analyses
library(rstatix) # 
library(cowplot) # plot_grid function
library(janitor) # clean column names
library(corrplot) # correlation matrix
library(ggbiplot) # PCA ordination plot
library(dunn.test) # Dunn's test (KW post hoc)
library(tidyverse)

# Site metadata
site_meta <- read_csv('raw_data/field_data/site_metadata.csv') %>% 
  mutate(severity = factor(severity, levels = c('unb', 'mod', 'high')))

# Field data for vegetation variables: transect
cover <- read_csv('raw_data/field_data/veg_transect_cover.csv') %>% clean_names()
cwd <- read_csv('raw_data/field_data/veg_transect_cwd.csv') %>% clean_names()
shrub <- read_csv('raw_data/field_data/veg_transect_shrub.csv') %>% clean_names()
tree <- read_csv('raw_data/field_data/veg_transect_tree.csv') %>% clean_names()

```

#### Tree density/survival

- **Tree survival:** percent live trees (%)
  - NOT included in GLM~veg~; used to justify fire severity categories
- **Live tree density:** number of live adult trees at both transects, divided by the transect area (number/hectare)

```{r include=FALSE}

# Tree survival
tree_survival <- tree %>%
  group_by(site) %>% 
  mutate(n = length(id)) %>% 
  filter(state != 'D') %>% 
  group_by(site, n) %>% 
  summarize(live = length(id)) %>% 
  mutate(tree_survival = live/n) %>% 
  select(site, tree_survival)

# Live tree density
live_tree_density <- tree %>% 
  filter(state !='D') %>% 
  group_by(site) %>% 
  summarize(n = length(id)) %>% 
  mutate(live_tree_density = n/1500*10000) %>% # density per hectare
  select(-n)

```

#### Cover

- **Shrub cover:** percent cover of live shrubs (%)
- **Forb/grass cover:** percent cover of live grass and live forbs combined (%)
- **Litter cover:** percent cover of litter, e.g. leaves, dead grass, small sticks (%)

```{r include=FALSE}

# Shrub cover
shrub_cover <- cover %>% 
  filter(type == 'Shrub') %>%
  group_by(site, transect, meter) %>% 
  summarize(quad_cover = sum(cover)) %>%
  group_by(site) %>% 
  summarize(shrub_cover = sum(quad_cover)/20)

# Forb/grass cover
forb_grass_cover <- cover %>% 
  filter(type == 'Dominant forb' | type == 'Forbs' | type == 'Grass') %>%
  group_by(site, transect, meter) %>% 
  summarize(quad_cover = sum(cover)) %>% 
  group_by(site) %>% 
  summarize(forb_grass_cover = sum(quad_cover)/20)

# Understory cover
understory_cover <- cover %>% 
  filter(type == 'Dominant forb' | type == 'Forbs' |
           type == 'Grass' | type == 'Shrub') %>%
  group_by(site, transect, meter) %>% 
  summarize(quad_cover = sum(cover)) %>%
  group_by(site) %>% 
  summarize(understory_cover = sum(quad_cover)/20)

# Leaf cover
leaf_cover <- cover %>% 
  filter(type == 'Soft loose litter') %>% 
  group_by(site, transect, meter) %>% 
  summarize(quad_cover = sum(cover, na.rm = TRUE)) %>% 
  group_by(site) %>% 
  summarize(leaf_cover = sum(quad_cover)/20)

# Litter cover
litter_cover <- cover %>% 
  filter(type == 'Soft loose litter' | type == 'Woody litter' |
           type == 'Soft rooted litter') %>% 
  group_by(site, transect, meter) %>% 
  summarize(quad_cover = sum(cover, na.rm = TRUE)) %>% 
  group_by(site) %>% 
  summarize(litter_cover = sum(quad_cover)/20)

```

#### Coarse Woody Debris

- **Coarse woody debris (CWD):** total volume per unit area (m^3^/hectare)
  - Calculated according to Waddell, 2002 using Smalian's and DeVries' formuals
  

```{r include=FALSE}

# Coarse woody debris
cwd_volume <- cwd %>%
  mutate(ind_volume = ((pi/8)*(diameter_small_cm^2 + diameter_large_cm^2)*length_m)/10000) %>% # volume of each piece of CWD in m^3 (Smalian's formula)
  mutate(ind_vol_weighted = ind_volume/length_m) %>% # volume of CWD, weighted by the length (DeVries' formula)
  group_by(site) %>% 
  summarize(sum = (sum(ind_vol_weighted))) %>% # sum of ind_vol_weighted (m^3/m)
  mutate(cwd_volume = (pi/200)*sum*10000) %>% # volume density per transect (m^3/m^2) (DeVries' formula), converted to m^3/hectare
  mutate(cwd_volume = replace_na(cwd_volume, 0)) %>% # sites with NAs had no CWD
  select(-sum)


# Soft CWD

soft_cwd_volume <- cwd %>%
  filter(decay>2) %>% 
  mutate(ind_volume = ((pi/8)*(diameter_small_cm^2 + diameter_large_cm^2)*length_m)/10000) %>% # volume of each piece of CWD in m^3 (Smalian's formula)
  mutate(ind_vol_weighted = ind_volume/length_m) %>% # volume of CWD, weighted by the length (DeVries' formula)
  group_by(site) %>% 
  summarize(sum = (sum(ind_vol_weighted))) %>% # sum of ind_vol_weighted (m^3/m)
  mutate(soft_cwd_volume = (pi/200)*sum*10000) %>% # volume density per transect (m^3/m^2) (DeVries' formula), converted to m^3/hectare
  mutate(soft_cwd_volume = replace_na(soft_cwd_volume, 0)) %>% # sites with NAs had no CWD
  select(-sum)

```

```{r include=FALSE}

veg_metrics <- full_join(site_meta[2:4], shrub_cover) %>% 
  full_join(forb_grass_cover) %>% 
  full_join(litter_cover) %>% 
  full_join(live_tree_density) %>%
  full_join(soft_cwd_volume) %>% 
  full_join(tree_survival) %>%
  arrange(severity)
veg_metrics[is.na(veg_metrics)] <- 0

```

<br>

### Comparison by fire severity

```{r}

veg <- pivot_longer(veg_metrics, cols=c(4:8), names_to = 'metric')

# Kruskal-Wallis
for (i in unique(veg$metric)) {
  dat <- filter(veg, metric==i)
  kw <- kruskal_test(dat, value~severity)
  kw_bonf <- p.adjust(kw$p, method='bonferroni', n=5)
  kw_eff <- kruskal_effsize(dat, value~severity)
  print(paste('Metric:',i))
  print(paste('H =',kw$statistic,'; p =',kw_bonf,'; effect size =',kw_eff$effsize,'; magnitude:', kw_eff$magnitude))
}

```

```{r eval=FALSE, include=FALSE}

# Post-hoc tests

veg %>% 
  filter(metric=='live_tree_density') %>% 
  dunn_test(value~severity, p.adjust.method='bonferroni')

veg %>% 
  filter(metric=='litter_cover') %>% 
  dunn_test(value~severity, p.adjust.method='bonferroni')

veg %>% 
  filter(metric=='shrub_cover') %>% 
  dunn_test(value~severity, p.adjust.method='bonferroni')

```

- Significant with a large effect size:
  - percent shrub cover
  - percent litter cover
  - live tree density

<br>

### Correlations between metrics

```{r}

# Correlations between metrics
pairs(~ live_tree_density + litter_cover + shrub_cover +
        forb_grass_cover + soft_cwd_volume,
      data=veg_metrics)

# Create a correlation matrix
veg_cor <- veg_metrics[4:8] %>% 
  cor(., method='spearman') # create correlation matrix

# Visualize the correlation matrix

corrplot(veg_cor,
         type = 'upper',
         order = 'hclust', # order variables by cluster
         tl.col = 'black',
         tl.cex = 0.8)

# Rank correlations

as.data.frame(veg_cor) %>% 
  mutate(variable1 = rownames(veg_cor)) %>%
  gather(variable2, cor_strength, 1:5) %>% 
  filter(cor_strength != 1) %>%
  arrange(-cor_strength) %>% 
  distinct(cor_strength, .keep_all = TRUE)

```

- Unsurprisingly, the three variables that changed with fire severity are highly correlated.

<br>

### PCA

```{r fig.width=4, fig.height=3}

# PCA of fire variables

veg_fire <- veg_metrics %>% 
  select(site, severity, shrub_cover, live_tree_density, litter_cover)

veg_pca <- prcomp(veg_fire[3:5],
                  scale = TRUE) # scales the data

colors <- c('darkgreen', 'orange', 'red')

ggbiplot(veg_pca, groups = veg_fire$severity) +
  scale_color_manual(name = 'Severity', values = colors) +
  xlim(-3, 3) +
  theme_classic()

pca_scores <- data.frame(veg_pca$x) %>% 
  bind_cols(veg_metrics[1:3]) %>% 
  select(site, severity, PC1) %>% 
  rename(pca_score = PC1)

summary(veg_pca) # PC1 describes 73.4% of the variance

# Write CSV
all_metrics <- full_join(veg_metrics, pca_scores)
write_csv(all_metrics,'output_data/08_veg_metrics.csv', col_names = TRUE)

```

```{r include=FALSE, eval=FALSE}

# Stats
# Kruskal Wallis, Dunn
all_metrics %>% kruskal_test(pca_score~severity)
all_metrics %>% kruskal_effsize(pca_score~severity)
all_metrics %>% dunn_test(pca_score~severity, p.adjust.method='bonferroni')

```

<br>

### Figures

```{r fig.width=6, fig.height=4}

cols <- c('#4AB793', '#DCA827', '#D85F2B')
sevs <- c('Unburned', 'Low-mod.\nseverity', 'High\nseverity')

set.seed(3)
shrub_plot <- filter(veg, metric=='shrub_cover') %>% 
  ggplot() +
  geom_jitter(aes(x=severity, y=value, color=severity, shape=severity), width=0.1, size=1.5) +
  geom_boxplot(aes(x=severity, y=value), fill=NA, outlier.shape=NA) +
  scale_color_manual(values=cols) +
  scale_shape_manual(values=c(19,17,15)) +
  theme_classic() +
  theme(legend.position = 'NA',
        axis.text.y = element_text(size = 9),
        axis.text.x = element_blank()) +
  labs(x='', y='Shrub cover\n(% cover)')

set.seed(2)
leaf_plot <- filter(veg, metric=='litter_cover') %>% 
  ggplot() +
  geom_jitter(aes(x=severity, y=value, color=severity, shape=severity), width=0.1, size=1.5) +
  geom_boxplot(aes(x=severity, y=value), fill=NA, outlier.shape=NA) +
  scale_color_manual(values=cols) +
  scale_shape_manual(values=c(19,17,15)) +
  scale_x_discrete(labels=sevs) +
  scale_y_continuous(limits=c(0,100)) +
  theme_classic() +
  theme(legend.position = 'NA',
        axis.text.y = element_text(size = 9)) +
  labs(x='', y='Litter cover\n(% cover)')

set.seed(15)
tree_plot <- filter(veg, metric=='live_tree_density') %>% 
  ggplot() +
  geom_jitter(aes(x=severity, y=value, color=severity, shape=severity), width=0.1, size=1.5) +
  geom_boxplot(aes(x=severity, y=value), fill=NA, outlier.shape=NA) +
  scale_color_manual(values=cols) +
  scale_shape_manual(values=c(19,17,15)) +
  theme_classic() +
  theme(legend.position = 'NA',
        axis.text.y = element_text(size = 9),
        axis.text.x = element_blank()) +
  scale_y_continuous(limits=c(0,600)) +
  labs(x='', y='Live tree density\n(trees/hectare)')

cwd_plot <- filter(veg, metric=='soft_cwd_volume') %>% 
  ggplot() +
  geom_jitter(aes(x=severity, y=value, color=severity, shape=severity), width=0.1, size=1.5) +
  geom_boxplot(aes(x=severity, y=value), fill=NA, outlier.shape=NA) +
  scale_color_manual(values=cols) +
  scale_shape_manual(values=c(19,17,15)) +
  theme_classic() +
  theme(legend.position = 'NA',
        axis.text.y = element_text(size = 9),
        axis.text.x = element_blank()) +
  labs(x='', y=expression(paste('Soft CWD (',m^2,'/ha)')))

ferb_plot <- filter(veg, metric=='forb_grass_cover') %>% 
  ggplot() +
  geom_jitter(aes(x=severity, y=value, color=severity, shape=severity), width=0.1, size=1.5) +
  geom_boxplot(aes(x=severity, y=value), fill=NA, outlier.shape=NA) +
  scale_color_manual(values=cols) +
  scale_shape_manual(values=c(19,17,15)) +
  scale_x_discrete(labels=sevs) +
  theme_classic() +
  theme(legend.position = 'NA',
        axis.text.y = element_text(size = 9),
        axis.title.y = element_text()) +
  scale_y_continuous(limits=c(0,30)) +
  labs(x='', y='Forb + grass cover\n(% cover)')

pca_plot <- ggplot(all_metrics) +
  geom_jitter(aes(x=severity, y=pca_score, color=severity, shape=severity),
              width=0.1, size=1.5) +
  geom_boxplot(aes(x=severity, y=pca_score), fill=NA, outlier.shape=NA) +
  scale_color_manual(values=cols) +
  scale_shape_manual(values=c(19,17,15)) +
  scale_x_discrete(labels=sevs) +
  theme_classic() +
  theme(legend.position = 'NA',
        axis.text.y = element_text(size = 9)) +
  labs(x='', y='Tree + Shrub + Litter\nPC1 (73.4% variance)')

plot_grid(tree_plot, shrub_plot, leaf_plot, pca_plot,
          ncol = 2, align = 'v',
          rel_heights = c(1,1.1))

ggsave('figs/08_fig5a.png', plot = last_plot(),
       width = 5.5, height = 4, units = 'in')

```

- **Figure 5a**
- Metrics that differed among fire severity categories, as well as PC1

```{r fig.height=4, fig.width=3}

plot_grid(cwd_plot, ferb_plot,
          ncol = 1, align = 'v',
          rel_heights = c(1,1.1))

ggsave('figs/08_fig5b.png', plot = last_plot(),
       width = 3, height = 4, units = 'in')

```

- **Figure 5b**
- Metrics that were similar among fire severity categories

```{r eval=FALSE, include=FALSE}
rmarkdown::render('code/08_vegetation_metrics.Rmd', output_file = '../docs/08_vegetation_metrics.html')
```
