==> code/09_mvglm.Rmd <==
---
title: "Multivariate GLM"
author: "Kate Culhane"
output: html_document
editor_options: 
  chunk_output_type: console
---

```{r setup, include=FALSE}
knitr::opts_knit$set(root.dir = '..')
knitr::opts_chunk$set(include = FALSE, warning = FALSE, message = FALSE,
                      fig.width=7, fig.height=5)
```

<br>

```{r packages and data, include=FALSE}

# Load packages

library(tidyverse)
# library(ggbiplot)
# library(randomForest)
# library(packfor)
library(mvabund)
# library(corrplot)
library(vegan)

# Update functions

select <- dplyr::select

# Load data

matrix_df <- read_csv('output_data/species_matrix.csv') %>% 
  mutate(severity = factor(severity, levels = c('unb', 'mod', 'high')))

veg_metrics <- read_csv('output_data/veg_metrics.csv') %>% 
  mutate(severity = factor(severity, levels = c('unb', 'mod', 'high')))

site_centroids <- read_csv('output_data/site_centroids.csv')

effort <- read_csv('output_data/effort.csv') %>% 
  mutate(offset = effort/300)

# diversity <- read_csv('output_data/diversity.csv') %>% 
#   mutate(severity = factor(severity, levels = c('unb', 'mod', 'high')))

```

```{r pcnm}

# Principle Components of Neighborhood Matrix (PCNM)

site_pcnm <- pcnm(dist(site_centroids[,2:3]))

pcnm_vectors <- as.data.frame(site_pcnm$vectors) %>% 
  cbind(site_centroids$sites_aea.site)%>% 
  rename(site = 'site_centroids$sites_aea.site',
         pcnm_1 = PCNM1) %>% 
  select(site, pcnm_1)

# All predictor variables

vars <- veg_metrics %>% 
  full_join(pcnm_vectors) %>% 
  select(-tree_survival)

```

### Multivariate GLM

Misc. resources:
http://eco-stats.blogspot.com/2012/03/introducing-mvabund-package-and-why.html
https://www.dropbox.com/s/unwd412a9wbkrl0/mvabundTute.R?dl=0
https://rdrr.io/cran/mvabund/man/anova.manyglm.html
http://www.seec.uct.ac.za/sites/default/files/image_tool/images/330/SEEC%20Toolbox%20mvabund.pdf

Whether to use anova() or summary(): https://stats.stackexchange.com/questions/59879/logistic-regression-anova-chi-square-test-vs-significance-of-coefficients-ano
Wald vs score statistic:https://stats.idre.ucla.edu/other/mult-pkg/faq/general/faqhow-are-the-likelihood-ratio-wald-and-lagrange-multiplier-score-tests-different-andor-similar/

**Hypothesis predictor variables:**

  - Shrub cover
  - Loose litter cover
  - Live tree density

```{r}

# Create mvabund object using species matrix
dat <- full_join(matrix_df, vars)
spp <- mvabund(dat[,3:13])

# Compare relationship between means and variances
meanvar.plot(spp) # looks like a linear increase...
# need to use appropriate GLM that assumes a similar mean-variance relationship
# negative binomial distribution is best for a linear mean-variance relationship

plot(data=dat, spp~severity)

```

<br>

### Severity effects

```{r}

# Test for differences by severity/block
# Does treatment have an effect on assemblage?
# anova() determines which predictors are the strongest contributors by sequentially adding them to the restricted (less variables) model (for test=score)
# (confirming the adonis results and nmds visualization)
glm <- manyglm(data=dat, spp~severity + pcnm_1,
               family='negative_binomial',
               offset = effort$offset) # account for trapping effort
plot(glm)
coef(glm)
an <- anova(glm,
            test='score', # use score statistic since means are small because of rare species, which decreases power in the Wald statistic
            cor.type='shrink', # accounts for correlation between variables
            nBoot=999)
an$table
sum <- summary(glm, test='score', cor.type='shrink', nBood=999)
sum$coefficients
sum$statistic

```

<br>

### Indicator species

```{r}

# Test univariate (species) differences by severity/block
# What are the indicator species?
an2 <- anova(glm, test='score', cor.type='shrink', nBoot=999, p.uni='adjusted')
an2
sum <- summary(glm, test='score', cor.type='shrink', nBood=999, p.uni='adjusted')


```

<br>

### Vegetation drivers

```{r}

# Test vegetation variables
# Which environmental variables are most strongly associated with assemblages?
# summary() determines which predictors are the strongest contributors when all other predictors have been fitted to the restricted (less variables) model (for test=score)
glm2 <- manyglm(data=dat, spp ~ pca_score + soft_cwd_volume + forb_grass_cover + pcnm_1,
               family='negative_binomial',
               offset=effort$offset)
plot(glm2)
coef(glm2)
glm2$AICsum
sum <- summary(glm2, test='score', cor.type='shrink', nBoot=999)
sum$coefficients
sum$statistic

best.r.sq(spp ~ dat$litter_cover + dat$shrub_cover + dat$live_tree_density +
            dat$soft_cwd_volume + dat$forb_grass_cover,
          n.xvars = 5, R2 = 'h')

# Create data frame with all possible variable combinations
veg <- c('soft_cwd_volume', 'forb_grass_cover', 'pca_score', 'pcnm_1')
cbind.fill <- function(...){
    nm <- list(...) 
    nm <- lapply(nm, as.matrix)
    n <- max(sapply(nm, nrow)) 
    do.call(cbind, lapply(nm, function (x) 
        rbind(x, matrix(,n-nrow(x),ncol(x))))) 
}
var_combos <- data.frame(cbind.fill(combn(veg,1), combn(veg,2), combn(veg,3), combn(veg,4)))

# Calculate AIC for each combination
model_aic <- data_frame(vars=character(),
                        K=numeric(),
                        aic=numeric())
for (i in 1:ncol(var_combos)) {
  vars <- levels(var_combos[,i])
  vars_dat <- as.matrix(select(dat,vars))
  K <- length(vars)
  aic <- manyglm(spp~vars_dat, family='negative_binomial', offset=effort$offset)$AICsum
  new_combo <- data_frame(vars=c(paste(vars,collapse=" + ")),
                          K=K,
                          aic=aic)
  model_aic <- rbind(model_aic, new_combo)
}

aic_arrange <- model_aic %>% 
  arrange(aic)

library(qpcR)
weights <- akaike.weights(aic_arrange$aic)

model_selection <- cbind(aic_arrange,
                         delta_aic=weights$deltaAIC,
                         weight=weights$weights) %>% 
  mutate(aic=round(aic,2),
         delta_aic=round(delta_aic,2),
         weight=round(weight,2))

# Final model
glm2 <- manyglm(data=dat, spp ~ pca_score + soft_cwd_volume,
               family='negative_binomial',
               offset=effort$offset)
plot(glm2)
coef(glm2)
sum <- summary(glm2, test='score', cor.type='shrink', nBoot=999)
sum$coefficients
sum$statistic

```

<br>

### Traits

Conceptually similar to fourth-corner analysis: https://rpubs.com/dwarton/68823
To emulate: https://royalsocietypublishing.org/doi/10.1098/rsos.200076#d3e468

```{r}

# Data
trait_new <- read.csv('raw_data/lit_data/mammal_traits_NEW.csv') %>% 
  arrange(species)
dat2 <- as.data.frame(dat)
L_dat <- dat2[c(3:13)] %>% 
  mutate_each(funs(if(is.numeric(.)) as.integer(.) else .))
rownames(L_dat) <- dat$site
R_dat <- dat2[c(2,19:20)]
rownames(R_dat) <- dat$site
traits <- trait_new[c(3:6)]
rownames(traits) <- trait_new$species

# Fourth corner

fc_glm <- traitglm(L_dat, R_dat[-1], traits[-1],
                   composition=TRUE, 
                   method='manyglm', family='negative.binomial')

plot(fc_glm)
qqnorm(residuals(fc_glm)); abline(c(0,1),col="red")

fc_anova <- anova(fc_glm)
fc_anova

fc <- traitglm(L_dat, R_dat[-1], traits[-1],
               composition=TRUE, # decreases heteroscedasticity
               method='glm1path', family='negative.binomial')

plot(fc)
qqnorm(residuals(fc)); abline(c(0,1),col="red")

fourth <- as_tibble(cbind(trait=rownames(fc$fourth),fc$fourth)) %>% 
  pivot_longer(cols=2:3,
               names_to='veg') %>% 
  separate(trait, sep=5, into=c('group','trait')) %>% 
  mutate(
    value=as.numeric(value),
    vegetation=case_when(
      veg=='pca_score'~'PC1[T+S+L]',
      veg=='soft_cwd_volume'~'Soft CWD'
    ),
    vegetation=factor(vegetation, levels=c('PC1[T+S+L]', 'Soft CWD')),
    group=case_when(
      group=='nests'~'Nesting habit',
      group=='guild'~'Feeding guild',
      group=='forag'~'Foraging mode'
    ),
    group=factor(group,levels=c('Feeding guild','Foraging mode','Nesting habit')),
    trait=factor(trait, levels=c('Insectivore','Omnivore','Herbivore',
                                 'Ground','Scansorial','Arboreal',
                                 'Burrow','Hollow','Tree'))
    )

fourth %>% 
  ggplot(aes(x=vegetation, y=trait, fill=value)) +
  geom_tile() +
  scale_fill_gradient2(low='darkred', high='darkblue', limits=c(-0.4,0.4)) +
  scale_x_discrete(labels=parse(text=c('PC1[T+S+L]','Soft~CWD'))) +
  facet_grid(rows=vars(group), scales='free_y', space='free', switch='y') +
  theme_minimal() +
  theme(panel.border = element_rect(fill=NA),
        panel.spacing = unit(0,'cm'),
        axis.text.x = element_text(angle=35, vjust=1.3, hjust=1.2,
                                   face='bold', color='black'),
        axis.ticks = element_line(),
        strip.placement = 'outside',
        strip.background.y = element_rect(fill=NA),
        strip.text = element_text(face='bold'),
        legend.key.height = unit(0.5,'in'),
        axis.title = element_text(size=10),
        legend.key.width = unit(0.1,'in'),
        legend.title = element_blank(),
        legend.box.spacing = unit(0.05,'in'),
        panel.grid = element_blank()) +
  labs(x='Vegetation variable', y='Small mammal trait')

ggsave('figs/fig_trait.png', plot = last_plot(),
       width = 2.5, height = 3.7, units = 'in')

```


```{r}

data(antTraits)

ft=traitglm(antTraits$abund,antTraits$env,antTraits$traits,method="manyglm")
ft$fourth #print fourth corner terms

# for a pretty picture of fourth corner coefficients, uncomment the following lines:
# library(lattice)
# a        = max( abs(ft$fourth.corner) )
# colort   = colorRampPalette(c("blue","white","red")) 
# plot.4th = levelplot(t(as.matrix(ft$fourth.corner)), xlab="Environmental Variables",
#   ylab="Species traits", col.regions=colort(100), at=seq(-a, a, length=100),
#   scales = list( x= list(rot = 45)))
# print(plot.4th)

plot(ft) # for a Dunn-smyth residual plot
qqnorm(residuals(ft)); abline(c(0,1),col="red") # for a normal quantile plot.

# predict to the first five sites
predict(ft,newR=antTraits$env[1:5,])

# refit using LASSO and less variables, including row effects and only two interaction terms:
ft1=traitglm(antTraits$abund,antTraits$env[,3:4],antTraits$traits[,c(1,3)],
      formula=~Shrub.cover:Femur.length+Shrub.cover:Pilosity,composition=TRUE,method="glm1path")
ft1$fourth #notice LASSO penalty has one interaction to zero

```



```{r eval=FALSE}
rmarkdown::render('code/09_mvglm.Rmd', output_file = '../docs/09_mvglm.html')
```
