==> code/10_mvglm.Rmd <==
---
title: "Multivariate GLM"
author: "Kate Culhane"
output:
  html_document:
    toc: true
    toc_float: true
editor_options: 
  chunk_output_type: console
---

```{r setup, include=FALSE}
knitr::opts_knit$set(root.dir = '..')
knitr::opts_chunk$set(echo = FALSE, warning = FALSE, message = FALSE,
                      fig.height=3.5, fig.width=5)
```

<br>

```{r packages and data, include=FALSE}

# Packages
library(adespatial) # MEM analysis
library(spdep) # MEM analysis
library(mvabund) # for constructing multivariate abundance models (including fourth-corner)
library(qpcR) # for calculating AIC weights
library(tidyverse)

# Small mammal abundance data (site by species matrix)
matrix_df <- read_csv('output_data/02_species_matrix.csv') %>% 
  mutate(severity = factor(severity, levels = c('unb', 'mod', 'high')))

# Small mammal trapping effort
effort <- read_csv('output_data/02_effort.csv') %>% 
  mutate(offset = effort/300) # calculate offset to use in model

# Vegetation variables
veg_metrics <- read_csv('output_data/08_veg_metrics.csv') %>% 
  mutate(severity = factor(severity, levels = c('unb', 'mod', 'high'))) %>% 
  select(-tree_survival)

# Site coordinates
site_coords <- read_csv('output_data/03_site_coords.csv') %>% 
  full_join(effort[1:2]) %>% 
  mutate(severity = factor(severity, levels = c('unb', 'mod', 'high'))) %>% 
  arrange(severity)

# NMDS scores
nmds_scores <- read_csv('output_data/09_nmds_scores.csv') %>% 
  mutate(severity = factor(group, levels = c('unb', 'mod', 'high'))) %>% 
  select(-group)

# Small mammal traits
traits <- read_csv('raw_data/lit_data/mammal_traits.csv') %>% 
  arrange(species)

# Create mvabund object using the site by species matrix and vegetation variables
dat <- full_join(matrix_df, veg_metrics)
spp <- mvabund(dat[,3:13])

```

## Moran's Eigenvector Maps (MEM)

**Purpose:** Create spatial filters that remove the spatial autocorrelation from the residuals of the community structure GLMs to meet assumption of independence

**Resources for MEM analysis:**

- [Bauman et al. 2018:](https://doi.org/10.1002/ecy.2469) Optimizing the choice of a spatial weighting matrix in eigenvector-based methods
- [Bauman et al. 2018, Appendix S3:](https://esajournals.onlinelibrary.wiley.com/action/downloadSupplement?doi=10.1002%2Fecy.2469&file=ecy2469-sup-0003-AppendixS3.pdf) Tutorial to optimize the selection of a spatial weighting matrix and/or the selection of a subset of spatial predictors in eigenvector-mapping methods

**Spatial weighting matrix (SWM)**

  - Describes the relations among a set of sampled sites
  - Calculated as the element-wise product of a connectivity matrix (B) and weighting matrix (A)
- *Connectivity matrix (B)*
  - Graph-based methods: Delaunay triangulation, Gabriel's graph, Relative neighbourhood graph, Minimum spanning tree
  - Distance based methods: Principal Coordinates of Neighbour Matrices (PCNM)
- *Weighting matrix (A)*

**SWM selection**

1. Determine candidate matrix B
  - Visualize the connection scheme and compare to realistic landscape connectivity
  - *Considering that it's highly unlikely that smamms traveled between the farthest sites (or across the river), we ruled out Delauney triangulation*
2. Determine candidate matrix A
  - View plot connectivity vs. distance with different functions (see Bauman et al. 2018, Appendix S1, Fig. S2)
  - *We chose a weighting matrix that uses a linear function to indicate that connectivity with distance (smamms are less likely to travel between distant sites); see Dray et al. 2006*
  - *The PCNM method does not incorporate weighting matrix, so we also included weights = 'binary' option*
3. Select a candidate using forward selection with model residuals
  -   Goal: Optimize the selection of the spatial variables unexplained by severity
  -   Multivariate data requires use of forward selection
  -   *We used residuals from a multivariate GLM (spp abundances ~ severity)*

```{r}

# Explore SWM possibilities (opens interactive window)
# coords <- as.matrix(site_coords[2:3])
# listw.explore()

# Create candidate SWMs
swm_can <- listw.candidates(site_coords[2:3], # uses only site coordinates
                            nb = c('gab', 'rel', 'mst', # connectivity matrix: graph-based
                                   'pcnm'), # connectivity matrix: distance-based (PCNM)
                            weights = c('binary', 'flin')) # weight matrix: absent or linear

# Adjust significance threshold (alpha) based on number of candidate SWMs
# Uses Sidak p-value correction
alpha <- round(1 - (1 - 0.05)^(1/length(swm_can)), 4) 

# Create model residuals for selecting SWM
# Multivariate GLM: spp abundances ~ severity
glm <- manyglm(data=dat, spp ~ severity, # multivariate GLM
               composition=FALSE, # use absolute abundance data
               offset=effort$offset, # account for trapping effort
               family='negative_binomial') # distribution of abundance data
set.seed(1)
glm_res <- residuals(glm) # multivariate Dunn-Smyth residuals

# Selection using multivariate GLM residuals
swm_sel <- listw.select(glm_res, # response variable Dunn-Smyth residuals
                        swm_can, # candidate SWMs 
                        MEM.autocor = "positive", # select positive eigenvectors
                        method = 'FWD', # use forward selection
                        alpha = alpha, # calculated above
                        p.adjust = TRUE)
swm_sel$candidates

# Checking multiple seeds because Dunn-Smyth residuals involve random number generation
seeds <- c(1:50)
for (i in seeds) {
  set.seed(i)
  glm_res <- residuals(glm) # multivariate Dunn-Smyth residuals
  swm_sel <- listw.select(glm_res, # response variable (Dunn-Smyth residuals)
                        swm_can, # candidate SWMs 
                        MEM.autocor = "positive", # select positive eigenvectors
                        method = 'FWD', # use forward selection
                        alpha = alpha, # calculated above
                        p.adjust = TRUE)
  if (is.null(swm_sel$best)) {
    print(paste0('Seed = ', i, ': NONE SIGNIFICANT'))
  } else {
    print(paste0('Seed = ', i))
    print(swm_sel$candidates)
  }
}

```

<br>

## Multivariate GLMs

Resources for `mvabund`:

-   [Introduction from Warton](http://eco-stats.blogspot.com/2012/03/introducing-mvabund-package-and-why.html)
-   [Tutorial](https://www.dropbox.com/s/unwd412a9wbkrl0/mvabundTute.R?dl=0)
-   [Powerpoint about mvabund](http://www.seec.uct.ac.za/sites/default/files/image_tool/images/330/SEEC%20Toolbox%20mvabund.pdf) (video lecture is also useful)
-   [Differences between `anova` and `summary`](https://stats.stackexchange.com/questions/59879/logistic-regression-anova-chi-square-test-vs-significance-of-coefficients-ano)
-   [Wald vs. score test statistic](https://stats.idre.ucla.edu/other/mult-pkg/faq/general/faqhow-are-the-likelihood-ratio-wald-and-lagrange-multiplier-score-tests-different-andor-similar/)
-   [Introduction to trait analysis by Warton](https://rpubs.com/dwarton/68823)

```{r eval=FALSE}

# Compare relationship between means and variances
meanvar.plot(spp) # looks like a linear increase...
# need to use appropriate GLM that assumes a similar mean-variance relationship --> negative binomial

# Exploratory plot
plot(data=dat, spp~severity)

```

<br>

## GLM~severity~

*Does treatment have an effect on assemblage?*

Testing for differences in multivariate community structure among the three fire severity categories

<br>

### Model

-   **Predictors:** severity

-   **Responses:** absolute abundances (`anova.manyglm` can't handle relative abundance)

    -   Trapping effort accounted for with an offset
    -   Assumes negative binomial distribution

```{r echo=TRUE}

glm_sev <- manyglm(data=dat, spp ~ severity,
                   composition=FALSE, # anova function (next) can't handle relative abundance
                   offset=effort$offset, # accounts for trapping effort
                   family='negative_binomial') # accounts for observed mean-variance relationship

plot(glm_sev) # check for model fit

```

<br>

### Significance (multivariate & univariate)

-   `anova.manyglm` determines which predictors are the strongest contributors by sequentially adding them to the restricted (less variables) model (for test=score)

    -   Used the score test to calculate test statistics since means are small because of rare species (small means decrease power when using the Wald test)
    -   Accounted for correlation between variables by shrinking the correlation matrix
    -   Calculated p-values using PIT-trap resampling with 999 bootstrap samples
    -   Calculated p-values for both multivariate and univariate (species-specific) responses

```{r echo=TRUE}

set.seed(2021)
an_sev <- anova(glm_sev,
                test='score', # allows for small means due to rare species
                cor.type='shrink', # accounts for correlation between variables
                resamp='pit.trap', # p-value calculated via PIT-trap resampling
                nBoot=999, # number of PIT-trap bootstrap iterations
                p.uni='adjusted') # also calculate univariate p-values
an_sev

```

<br>

## GLM~veg~

*Which environmental variables are most strongly associated with assemblages?*

Testing for differences in multivariate small mammal community structure according to the vegetation variables

<br>

### AIC model selection

**Candidate predictors:**

-   PC1~TSL~ (Live tree density + Shrub cover + Litter cover)
-   Soft coarse woody debris (CWD)
-   Forb/grass cover

```{r echo=FALSE}

# Create data frame with all possible variable combinations
veg <- c('soft_cwd_volume', 'forb_grass_cover', 'pca_score')
cbind.fill <- function(...){
    nm <- list(...) 
    nm <- lapply(nm, as.matrix)
    n <- max(sapply(nm, nrow)) 
    do.call(cbind, lapply(nm, function (x) 
        rbind(x, matrix(,n-nrow(x),ncol(x))))) 
}
var_combos <- data.frame(cbind.fill(combn(veg,1), combn(veg,2), combn(veg,3)))

# Calculate AIC for each combination
model_aic <- data.frame(vars=character(),
                        K=numeric(),
                        aic=numeric())
for (i in 1:ncol(var_combos)) {
  vars <- unique(na.omit(rev(var_combos[,i])))
  vars_dat <- as.matrix(select(dat,vars))
  k <- length(vars)
  aic <- manyglm(spp~vars_dat,
                 composition=FALSE,
                 offset=effort$offset,
                 family='negative_binomial')$AICsum
  new_combo <- data.frame(vars=c(paste(vars,collapse=" + ")),
                          k=k,
                          aic=aic)
  model_aic <- rbind(model_aic, new_combo)
}

aic_arrange <- model_aic %>% 
  arrange(aic)

weights <- akaike.weights(aic_arrange$aic)

model_selection <- cbind(aic_arrange,
                         delta_aic=weights$deltaAIC,
                         weight=weights$weights) %>% 
  mutate(aic=round(aic,2),
         delta_aic=round(delta_aic,2),
         weight=round(weight,2))

model_selection

```

<br>

### Final model

-   **Final predictors:** PC1~TSL~ + Soft CWD

-   **Responses:** absolute abundances

    -   Trapping effort accounted for with an offset
    -   Assumes negative binomial distribution

```{r echo=TRUE}

glm_veg <- manyglm(data=dat, spp ~ pca_score + soft_cwd_volume,
                   composition=FALSE, # absolute abundance
                   offset=effort$offset, # trapping effort
                   family='negative_binomial') # data distribution

plot(glm_veg) # check for model fit

```

<br>

### Significance

-   `summary` determines which predictors are the strongest contributors when all other predictors have been fitted to the restricted (less variables) model (true for for test=score)

    -   Used the score test to calculate test statistics
    -   Accounted for correlation between variables by shrinking the correlation matrix
    -   Calculated p-values using PIT-trap resampling with 999 bootstrap samples

```{r echo=TRUE}

set.seed(2021)
sum_veg <- summary(glm_veg,
                   test='score', # allows for small means due to rare species
                   cor.type='shrink', # accounts for correlation between variables
                   resamp='pit.trap', # p-value calculated via PIT-trap resampling
                   nBoot=999) # number of PIT-trap bootstrap iterations
sum_veg

```

<br>

## GLM~trait~

Fourth-corner analysis (trait by environment interactions)

```{r include=FALSE}

# Data wrangling
dat2 <- as.data.frame(dat)
traits2 <- as.data.frame(traits)
L_dat <- dat2[c(3:13)] %>% # abundance
  mutate_each(funs(if(is.numeric(.)) as.integer(.) else .))
rownames(L_dat) <- dat$site
R_dat <- dat2[c(2,19:20)] # environment
rownames(R_dat) <- dat$site
Q_dat <- traits2[c(6:9)] # traits
rownames(Q_dat) <- traits$species

```

### Model

-   **Predictors:** interactions between vegetation variables (PC1~TSL~ + Soft CWD) and small mammal traits (feeding guild + foraging mode + nesting habit)

-   **Responses:** relative small mammal abundances

    -   No option for including an offset to account for trapping effort, so used relative abundance data instead
    -   Using relative instead of absolute abundance improved model fit by decreasing heteroscedasticity of residuals
    -   Assumes negative binomial distribution

```{r echo=TRUE}

glm_trait <- traitglm(L_dat, R_dat[-1], Q_dat[-1],
                      composition=TRUE, # relative abundance
                      method='manyglm',
                      family='negative.binomial') # accounts for mean-variance relationship

plot(glm_trait) # check for model fit

```

<br>

### Significance

-   `anova` tests whether the environment-trait interactions predict community structure

    -   Used the score test to calculate test statistics
    -   No need to shrink the correlation matrix
    -   Calculated p-values using PIT-trap resampling with 999 bootstrap samples

```{r echo=TRUE}

set.seed(2021)
an_trait <- anova(glm_trait,
                  test='score', # allows for small means due to rare species
                  resamp='pit.trap', # p-value calculated via PIT-trap resampling
                  nBoot=999) # number of PIT-trap bootstrap iterations
an_trait$table

```

<br>

### LASSO penalty model selection

-   Uses the LASSO penalty to do model selection (removes interaction terms that do not decrease BIC)
-   Cannot use `anova` with this method
-   Shows which environment-trait interactions are most important for determining community structure

```{r echo=TRUE}

# LASSO penalty
set.seed(2021)
glm_fc <- traitglm(L_dat, R_dat[-1], Q_dat[-1],
                   composition=TRUE, # relative abundance
                   method='glm1path', # LASSO penalty
                   family='negative.binomial') # accounts for mean-variance relationship

plot(glm_fc) # check model fit

```

<br>

### Heat map

Trait-environment interaction coefficient estimates

```{r echo=FALSE, fig.height=4, fig.width=3}

fourth <- as_tibble(cbind(trait=rownames(glm_fc$fourth),glm_fc$fourth)) %>% 
  pivot_longer(cols=2:3,
               names_to='veg') %>% 
  separate(trait, sep=5, into=c('group','trait')) %>% 
  mutate(
    value=as.numeric(value),
    vegetation=case_when(
      veg=='pca_score'~'PC1[T+S+L]',
      veg=='soft_cwd_volume'~'Soft CWD'
    ),
    vegetation=factor(vegetation, levels=c('PC1[T+S+L]', 'Soft CWD')),
    group=case_when(
      group=='nests'~'Nesting habit',
      group=='guild'~'Feeding guild',
      group=='forag'~'Foraging mode'
    ),
    group=factor(group,levels=c('Feeding guild','Foraging mode','Nesting habit')),
    trait=factor(trait, levels=c('Insectivore','Omnivore','Herbivore',
                                 'Ground','Scansorial','Arboreal',
                                 'Burrow','Hollow','Tree')),
    label=case_when(
      value==0~NA_real_,
      TRUE~round(value,2)
    )
    )

fourth %>% 
  ggplot(aes(x=vegetation, y=trait)) +
  geom_tile(aes(fill=value)) +
  scale_fill_gradient2(low='darkblue', high='darkred', limits=c(-0.4,0.4)) +
  geom_text(aes(label=label,
                color=value>0.2 | value<c(-0.2)), size=2.5) +
  scale_color_manual(guide = FALSE, values = c("black", "white")) +
  scale_x_discrete(labels=parse(text=c('PC1[T+S+L]','Soft~CWD'))) +
  facet_grid(rows=vars(group), scales='free_y', space='free', switch='y') +
  theme_minimal() +
  theme(panel.border = element_rect(fill=NA),
        panel.spacing = unit(0,'cm'),
        axis.text.x = element_text(angle=35, vjust=1.3, hjust=1.2,
                                   face='bold', color='black'),
        axis.ticks = element_line(),
        strip.placement = 'outside',
        strip.background.y = element_rect(fill=NA),
        strip.text = element_text(face='bold'),
        legend.key.height = unit(0.5,'in'),
        axis.title = element_text(size=10),
        legend.key.width = unit(0.1,'in'),
        legend.title = element_blank(),
        legend.box.spacing = unit(0.05,'in'),
        panel.grid = element_blank()) +
  labs(x='Vegetation variable', y='Small mammal trait')

ggsave('figs/10_fig6.png', plot = last_plot(),
       width = 2.5, height = 3.7, units = 'in')

```

```{r eval=FALSE, include=FALSE}
rmarkdown::render('code/10_mvglm.Rmd', output_file = '../docs/10_mvglm.html')
```
