==> code/11_rare_species_excluded.Rmd <==
---
title: "Analyses excluding rare species"
author: "Kate Culhane"
output:
  html_document:
    toc: true
    toc_float: true
editor_options: 
  chunk_output_type: console
---

```{r setup, include=FALSE}
knitr::opts_knit$set(root.dir = '..')
knitr::opts_chunk$set(echo = FALSE, warning = FALSE, message = FALSE,
                      fig.width=5, fig.height=3)
```

<br>

Major analyses, excluding rare species (<5 individuals captured):

- Dusky-footed woodrat (NEFU)
- Yellow pine chipmunk (NEAM)
- Northern flying squirrel (GLSA)
- Pinyon mouse (PETR)
- Western harvest mouse (REME)

<br>

## Data wrangling

```{r include=FALSE}

# Packages
library(kableExtra) # making pretty tables
library(rstatix) # KW test, KW eff
library(dunn.test) # Dunn's test (KW post hoc)
library(effectsize) # 
library(vegan) # Community analyses (diversity, NMDS, adonis, PCNM)
library(goeveg) # 
library(pairwiseAdonis) # post hoc adonis
library(ggvegan) # making pretty NMDS plot
library(ggrepel) # making pretty NMDS plot
library(mvabund) # multivariate models
library(qpcR) # for calculating AIC weights
library(tidyverse)

# Small mammal capture data
smamms_clean <- read_csv('output_data/01_smamms_clean.csv') %>%
  mutate(species = case_when( 
    species=='NEFU' ~ 'SPRU', # remove rare species and code as sprung (0.5 trapping effort)
    species=='NEAM' ~ 'SPRU',
    species=='GLSA' ~ 'SPRU',
    species=='PETR' ~ 'SPRU',
    species=='REME' ~ 'SPRU',
    TRUE ~ species
  ))

# Site metadata
site_meta <- read_csv('raw_data/field_data/site_metadata.csv') %>%
    mutate(severity = factor(severity, levels = c('unb', 'mod', 'high')))

# Small mammal trait data
traits <- read_csv('raw_data/lit_data/mammal_traits.csv') %>% 
  arrange(species) %>% 
  filter(species!='NEFU', species!='NEAM', species!='GLSA', species!='PETR', species!='REME') %>% 
  mutate(guild=factor(guild),
         nests=factor(nests),
         forag=factor(forag))

# Vegetation variables 
veg_metrics <- read_csv('output_data/08_veg_metrics.csv') %>% 
  mutate(severity = factor(severity, levels = c('unb', 'mod', 'high')))

# Site coordinates
site_coords <- read_csv('output_data/03_site_coords.csv') %>% 
  full_join(effort[1:2]) %>% 
  mutate(severity = factor(severity, levels = c('unb', 'mod', 'high'))) %>% 
  arrange(severity)

```

```{r include=FALSE}

# Data wrangling

# Create full smamm data frame with site and species information
captures <- full_join(smamms_clean, site_meta) %>%
  left_join(traits) %>% 
  select(site, severity, date, day, trap, indID, species, binomial, binomial_short, common_name, family, age, recap, weight, comments)

# Create data frame with one row per unique individual
smamms <- captures %>% 
  filter(species != 'MISS' & species != 'SPRU') %>% # remove SPRU and MISS
  group_by(indID) %>% 
  arrange(-day) %>% 
  filter(row_number()==1) %>% # filter by unique animals (only keep record from last capture)
  ungroup() %>% 
  select(-day)

# Data frame of trapping effort at each site
effort <- captures %>% 
  filter(species == 'SPRU' | species == 'MISS') %>%
  mutate(usage = case_when(
    species == 'SPRU' ~ 0.5, # sprung traps coded as 0.5 usage
    species == 'MISS' ~ 1 # missing traps coded as 1 usage (to be subtracted)
  )) %>% 
  group_by(site, severity) %>% 
  summarize(sprung = sum(usage)) %>% 
  mutate(effort =300 - sprung) %>% 
  select(site, severity, effort) %>% 
  arrange(severity) %>% 
  mutate(offset = effort/300) # for use in multivariate GLMs

# Species matrix for use in vegan
matrix_df <- smamms %>% 
  group_by(site, severity, species) %>% 
  tally() %>% # count individuals per species per site
  spread(species, n) %>% # re-format into matrix
  ungroup() %>% 
  replace(., is.na(.), 0) %>% # replace NAs with 0
  arrange(severity)

# Species matrix adjusted for trapping effort
matrix_adj_df <- smamms %>% 
  group_by(site, severity, species) %>% 
  tally() %>% # count individuals per species per site
  full_join(effort, by = c('site', 'severity')) %>% 
  mutate(n_adj = n/effort*300) %>% # add column for n adjusted for effort
  select(site, species, n_adj)%>% 
  spread(species, n_adj) %>% # re-format into matrix
  ungroup() %>% 
  replace(., is.na(.), 0) %>% # replace NAs with 0
  arrange(severity)

```

**Site by species matrix**

```{r}

kable(matrix_df) %>% 
  kable_styling(bootstrap_options = c('striped', 'condensed'),
                full_width = FALSE,
                position = 'left')

```

<br>

## Abundance

```{r include=FALSE}

# Vectors for plotting
cols <- c('#4AB793', '#DCA827', '#D85F2B')
xlabs <- c('Unburned', 'Moderate severity', 'High severity') 

# Calculate total abundance summed across all species at each site (raw and adjusted)
abundance <- smamms %>% 
  group_by(site, severity) %>% 
  tally() %>% # count individuals per site
  full_join(effort) %>% 
  mutate(n_adj = n/effort*300) %>% # add column adjusted for effort
  ungroup() %>% 
  full_join(site_meta[2:4]) %>% 
  mutate(offset=effort/300)

```

### Total abundance

```{r}

# Plot
ggplot(abundance, aes(x = severity, y = n_adj)) +
  geom_jitter(aes(color = severity, shape = severity), width = 0.2, size = 2) +
  scale_color_manual(values = cols) +
  scale_shape_manual(values=c(19,17,15)) +
  geom_boxplot(outlier.shape=NA, fill=NA) +
  theme_classic() +
  labs(x = 'Fire Severity',
       y = 'Total small mammal abundance') +
  theme(legend.position = 'NA') +
  scale_x_discrete(labels = xlabs)

# Stats
kruskal.test(n_adj~severity, data=abundance)
print.data.frame(kruskal_effsize(n_adj~severity, data=abundance))

```

<br>

### Biomass

```{r include=FALSE}

# Average mass for each unique individual
capture_mass <- captures %>% 
  select(site, severity, indID, binomial, weight, age) %>% 
  filter(!is.na(indID), # filter for captures only
         age == 'A') %>% # filter for only adult animals
  group_by(indID, site, severity, binomial) %>% 
  summarize(avg_mass = mean(weight, na.rm = TRUE)) %>% 
  ungroup()

# Average mass for each species
species_mass <- capture_mass %>% 
  group_by(binomial) %>% 
  summarize(spp_avg = mean(avg_mass, na.rm = TRUE))

# Average mass for each unique individual, with NAs filled in with species averages
capture_biomass <- full_join(capture_mass, species_mass, by = 'binomial') %>% 
  mutate(mass = case_when(
    !is.na(avg_mass) ~ avg_mass,
    is.na(avg_mass) ~ spp_avg
  )) %>% 
  select(site, severity, binomial, mass)

# Total biomass at each site
biomass <- capture_biomass %>% 
  group_by(site, severity) %>% 
  summarize(biomass = sum(mass)) %>% 
  ungroup()

```


```{r}

# Plot
ggplot(biomass, aes(x = severity, y = biomass)) +
  geom_jitter(aes(color = severity, shape = severity), width = 0.2, size = 2) +
  scale_color_manual(values = cols) +
  scale_shape_manual(values=c(19,17,15)) +
  stat_summary(fun.data = mean_se, geom = 'errorbar', width = 0.2, size = 0.7) +
  stat_summary(fun = 'mean', size = 2, geom = 'point') +
  theme_classic() +
  labs(x = 'Fire Severity',
       y = 'Total small mammal biomass') +
  theme(legend.position = 'NA') +
  scale_x_discrete(labels = xlabs)

# Stats
kruskal.test(biomass~severity, data=biomass)
print.data.frame(kruskal_effsize(biomass~severity, data=biomass))

```

## Diversity

```{r include=FALSE}

# Transform data frames into matrices
matrix <- as.matrix(matrix_df[-c(1,2)])
rownames(matrix) <- matrix_df$site
matrix_adj <- as.matrix(matrix_adj_df[-c(1,2)])
rownames(matrix_adj) <- matrix_adj_df$site

# Vectors for plotting
cols <- c('darkgreen', 'orange', 'red')

```

### Per-site richness

*Individual-based rarefaction estimates*

```{r, include=FALSE}

# Calculate richness for each site using rarefaction
raremax <- min(rowSums(matrix))
raremax # minimum sample count = 5: rarefy using this value
rarefied_richness <- rarefy(matrix, sample = raremax)

# Make a dataframe with site, severity, and richness
richness <- tibble(site = matrix_df$site, severity = matrix_df$severity, rarefied_richness)

```

```{r}

# Plot
ggplot(richness, aes(x = severity, y = rarefied_richness)) +
  geom_jitter(aes(color = severity)) +
  scale_color_manual(values = cols) +
  stat_summary(fun.data = mean_se, geom = "crossbar", width = .5) +
  theme_classic() +
  labs(x = 'Fire Severity',
       y = 'Rarefied Richness') +
  theme(legend.position = 'NA') +
  scale_x_discrete(labels = xlabs)

# Stats
aov <- aov(rarefied_richness~severity, data=richness)
summary(aov)
richness %>% ungroup() %>% tukey_hsd(rarefied_richness~severity) %>% print.data.frame()
cohens_f(aov)

```

<br>

### Pielou's Evenness (J)

```{r include=FALSE}

# Calculate Pielou's Evenness
J <- diversity(matrix_adj)/log(specnumber(matrix_adj)) # H / log (number of species)
diversity <- cbind(richness, J)
diversity$J[is.nan(diversity$J)] <- NA

```

```{r}

# Plot
ggplot(diversity, aes(x = severity, y = J)) +
  geom_jitter(aes(color = severity)) +
  scale_color_manual(values = cols) +
  stat_summary(fun.data = mean_se, geom = "crossbar", width = .5) +
  theme_classic() +
  labs(x = 'Fire Severity',
       y = 'Pielou\'s Evenness (J)') +
  theme(legend.position = 'NA') +
  scale_x_discrete(labels = xlabs)

# Stats
aov <- aov(J~severity, data=diversity)
summary(aov)
diversity %>% ungroup() %>% tukey_hsd(J~severity) %>% print.data.frame()
cohens_f(aov)

```

<br>

## NMDS

```{r include=FALSE}

# Transform matrices
matrix_hell <- decostand(matrix, 'hellinger')
matrix_sqrt <- sqrt(matrix)

# Severity data frame
sevs <- select(matrix_df, site, severity)

```

```{r include=FALSE}

# NMDS with relative abundance data (2 dimensions)
set.seed(17)
ord <- metaMDS(matrix_adj,
               distance = 'bray',
               autotransform = FALSE,
               k = 2,
               try=50)
ord$stress # 0.08

```

```{r, fig.height=4, results='hide',fig.keep='all', fig.show="hold", out.width="50%"}

# Stress plot
stressplot(ord)

# Scree plot to check stress per number of dimensions
dimcheckMDS(matrix_hell, distance = "bray", autotransform = FALSE, k = 10, trymax = 20)

```

```{r}

cols <- c('darkgreen', 'orange', 'red')
par(mar = c(4,4,1,1))
plot(ord, display = 'species', type = 'n')
points(ord, display = 'sites',
       cex = 1.5, pch = 19, col = cols[sevs$severity])
ordiellipse(ord, groups = sevs$severity,
            label = TRUE, col = cols, lwd = 2)
text(ord, display = 'species',
     cex = 0.6)

```

<br>

### Adonis

```{r include=FALSE}

# Attach severity data to species matrix
matrix_hell
attach(sevs)

# Compare community similarity across fire severities using adonis
smamm_ado <- adonis(matrix_hell ~ severity,
                    permuatations = 999,
                    method = 'bray')

# Pairwise Adonis
smamm_ado_pw <- pairwise.adonis(matrix_hell, severity,
                                sim.method = 'bray',
                                p.adjust.m = 'bonferroni')

```

```{r}

smamm_ado
smamm_ado_pw

```

<br>

### Habitat drivers

```{r include=FALSE}

# Data frame with only the variables from the model
model_vars <- veg_metrics %>% 
  select(site, severity,
         litter_cover, live_tree_density, soft_cwd_volume,
         shrub_cover, forb_grass_cover, pca_score)

# Fit environmental variables
fit <- envfit(ord, model_vars[,3:8], permutations = 1000)

```

```{r}

# Plot
par(mar = c(4,4,1,1))
plot(ord, diplay = 'sites', type = 'n')
points(ord, display = 'sites',
       cex = 1, pch = 19, col = cols[sevs$severity])
ordiellipse(ord, groups = sevs$severity, # ellipse with std error around centroids
            label = FALSE, col = cols, lwd = 2)
plot(fit,
     col = 'black', lwd = 3, cex = 0.8)
text(ord, display = 'species',
     cex = 0.7, col = 'grey30')

fit

```

<br>

## Multivariate GLMS

```{r include=FALSE}

# Create mvabund object using the site by species matrix and vegetation variables
dat <- full_join(matrix_df, veg_metrics)
spp <- mvabund(dat[,3:8])

```

```{r include=FALSE, eval=FALSE}

# Compare relationship between means and variances
meanvar.plot(spp) # looks like a linear increase
plot(data=dat, spp~severity)

```

<br>

### GLM~severity~

- **Predictors:** severity + PCNM
- **Responses:** absolute abundances

```{r include=FALSE}

# Model
glm_sev <- manyglm(data=dat, spp ~ severity,
                   composition=FALSE, # anova function (next) can't handle relative abundance
                   offset=effort$offset, # accounts for trapping effort
                   family='negative_binomial') # accounts for observed mean-variance relationship

plot(glm_sev) # check for model fit

# Univariate and multivariate significance
set.seed(2021)
an_sev <- anova(glm_sev,
                test='score', # allows for small means due to rare species
                cor.type='shrink', # accounts for correlation between variables
                resamp='pit.trap', # p-value calculated via PIT-trap resampling
                nBoot=999, # number of PIT-trap bootstrap iterations
                p.uni='adjusted') # calculates univariate and multivariate p-values
```

```{r}
an_sev
```

<br>

### GLM~veg~

**Candidate predictors:**

- PC1~TSL~ (Live tree density + Shrub cover + Litter cover)
- Soft coarse woody debris (CWD)
- Forb/grass cover 
- PCNM

**AIC selection:**

```{r echo=FALSE}

# Create data frame with all possible variable combinations
veg <- c('soft_cwd_volume', 'forb_grass_cover', 'pca_score')
cbind.fill <- function(...){
    nm <- list(...) 
    nm <- lapply(nm, as.matrix)
    n <- max(sapply(nm, nrow)) 
    do.call(cbind, lapply(nm, function (x) 
        rbind(x, matrix(,n-nrow(x),ncol(x))))) 
}
var_combos <- data.frame(cbind.fill(combn(veg,1), combn(veg,2), combn(veg,3)))

# Calculate AIC for each combination
model_aic <- data.frame(vars=character(),
                        K=numeric(),
                        aic=numeric())
for (i in 1:ncol(var_combos)) {
  vars <- unique(na.omit(rev(var_combos[,i])))
  vars_dat <- as.matrix(select(dat,vars))
  k <- length(vars)
  aic <- manyglm(spp~vars_dat,
                 composition=FALSE,
                 offset=effort$offset,
                 family='negative_binomial')$AICsum
  new_combo <- data.frame(vars=c(paste(vars,collapse=" + ")),
                          k=k,
                          aic=aic)
  model_aic <- rbind(model_aic, new_combo)
}

aic_arrange <- model_aic %>% 
  arrange(aic)

weights <- akaike.weights(aic_arrange$aic)

model_selection <- cbind(aic_arrange,
                         delta_aic=weights$deltaAIC,
                         weight=weights$weights) %>% 
  mutate(aic=round(aic,2),
         delta_aic=round(delta_aic,2),
         weight=round(weight,2))

model_selection

```

<br>

- **Final predictors:** PC1~TSL~ + Soft CWD
- **Responses:** absolute abundances

<br>

```{r include=FALSE}

# Model
glm_veg <- manyglm(data=dat, spp ~ pca_score + soft_cwd_volume,
                   composition=FALSE, # absolute abundance
                   offset=effort$offset, # trapping effort
                   family='negative_binomial') # data distribution

plot(glm_veg) # check for model fit

# Significance
set.seed(2021)
sum_veg <- summary(glm_veg,
                   test='score', # allows for small means due to rare species
                   cor.type='shrink', # accounts for correlation between variables
                   resamp='pit.trap', # p-value calculated via PIT-trap resampling
                   nBoot=999) # number of PIT-trap bootstrap iterations

```

```{r}
sum_veg
```

### GLM~trait~

```{r include=FALSE}

# Data wrangling
dat2 <- as.data.frame(dat)
traits2 <- as.data.frame(traits)
L_dat <- dat2[c(3:8)] %>% # abundance
  mutate_each(funs(if(is.numeric(.)) as.integer(.) else .))
rownames(L_dat) <- dat$site
R_dat <- dat2[c(2,14:15)] # environment
rownames(R_dat) <- dat$site
Q_dat <- traits2[c(6:9)] # traits
rownames(Q_dat) <- traits$species

```

- **Predictors:** interactions between vegetation variables (PC1~TSL~ + Soft CWD) and small mammal traits (feeding guild + foraging mode + nesting habit)
- **Responses:** relative small mammal abundances

```{r include=FALSE}

# Model
glm_trait <- traitglm(L_dat, R_dat[-1], Q_dat[-1],
                      composition=TRUE, # relative abundance
                      method='manyglm',
                      family='negative.binomial') # accounts for mean-variance relationship

plot(glm_trait) # check for model fit

# Significance
set.seed(2021)
an_trait <- anova(glm_trait,
                  test='score', # allows for small means due to rare species
                  resamp='pit.trap', # p-value calculated via PIT-trap resampling
                  nBoot=999) # number of PIT-trap bootstrap iterations

```

```{r}
an_trait
```

<br>

**LASSO penalty model selection**

```{r include=FALSE}

# Model
glm_fc <- traitglm(L_dat, R_dat[-1], Q_dat[-1],
                   composition=TRUE, # relative abundance
                   method='glm1path', # LASSO penalty
                   family='negative.binomial') # accounts for mean-variance relationship

plot(glm_fc) # check model fit

```

Heat map of interaction coefficient estimates

```{r fig.width=3, fig.height=4}

# Heat map
fourth <- as_tibble(cbind(trait=rownames(glm_fc$fourth),glm_fc$fourth)) %>% 
  pivot_longer(cols=2:3,
               names_to='veg') %>% 
  separate(trait, sep=5, into=c('group','trait')) %>% 
  mutate(
    value=as.numeric(value),
    vegetation=case_when(
      veg=='pca_score'~'PC1[T+S+L]',
      veg=='soft_cwd_volume'~'Soft CWD'
    ),
    vegetation=factor(vegetation, levels=c('PC1[T+S+L]', 'Soft CWD')),
    group=case_when(
      group=='nests'~'Nesting habit',
      group=='guild'~'Feeding guild',
      group=='forag'~'Foraging mode'
    ),
    group=factor(group,levels=c('Feeding guild','Foraging mode','Nesting habit')),
    trait=factor(trait, levels=c('Insectivore','Omnivore','Herbivore',
                                 'Ground','Scansorial','Arboreal',
                                 'Burrow','Hollow','Tree')),
    label=case_when(
      value==0~NA_real_,
      TRUE~round(value,2)
    )
    )

fourth %>% 
  ggplot(aes(x=vegetation, y=trait)) +
  geom_tile(aes(fill=value)) +
  scale_fill_gradient2(low='darkblue', high='darkred', limits=c(-0.4,0.4)) +
  geom_text(aes(label=label,
                color=value>0.2 | value<c(-0.2)), size=2.5) +
  scale_color_manual(guide = 'none', values = c("black", "white")) +
  scale_x_discrete(labels=parse(text=c('PC1[T+S+L]','Soft~CWD'))) +
  facet_grid(rows=vars(group), scales='free_y', space='free', switch='y') +
  theme_minimal() +
  theme(panel.border = element_rect(fill=NA),
        panel.spacing = unit(0,'cm'),
        axis.text.x = element_text(angle=35, vjust=1.3, hjust=1.2,
                                   face='bold', color='black'),
        axis.ticks = element_line(),
        strip.placement = 'outside',
        strip.background.y = element_rect(fill=NA),
        strip.text = element_text(face='bold'),
        legend.key.height = unit(0.5,'in'),
        axis.title = element_text(size=10),
        legend.key.width = unit(0.1,'in'),
        legend.title = element_blank(),
        legend.box.spacing = unit(0.05,'in'),
        panel.grid = element_blank()) +
  labs(x='Vegetation variable', y='Small mammal trait')

```

```{r, eval=FALSE}
rmarkdown::render('code/11_rare_species_excluded.Rmd', output_file = '../docs/11_rare_species_excluded.html')
```
