==> code/12_site_justification.Rmd <==
---
title: "Site Justification"
author: "Kate Culhane"
output: html_document
editor_options: 
  chunk_output_type: console
---

```{r setup, include=FALSE}
knitr::opts_knit$set(root.dir = '..')
knitr::opts_chunk$set(echo = FALSE, warning = FALSE, message = FALSE,
                      fig.width=5, fig.height=3)
```

```{r packages & data}

# Load packages

library(tidyverse)
library(cowplot) # plot_grid function
library(dunn.test)
library(ggbiplot)
library(rstatix)
library(ggpubr)

# Update functions

rename <- dplyr::rename
summarize <- dplyr::summarize
mutate <- dplyr::mutate

# Load data

site_meta <- read_csv('raw_data/field_data/site_metadata.csv')

spatial <- read_csv('output_data/spatial_metrics.csv')

veg <- read_csv('output_data/veg_metrics.csv')

crosswalk <- read_csv('raw_data/spatial_data/crosswalk.csv') # for converting between evt and cwhr

evt_2012 <- read_csv('output_data/EVT_2012.csv') %>% # pre-fire
  mutate(severity = factor(severity, levels = c('unb', 'mod', 'high'))) 
evt_12_meta <- read_csv('raw_data/spatial_data/EVT_2012.csv')
evt_2014 <- read_csv('output_data/EVT_2014.csv') %>% # post-fire
  mutate(severity = factor(severity, levels = c('unb', 'mod', 'high'))) 
evt_14_meta <- read_csv('raw_data/spatial_data/EVT_2014.csv')

evc_2012 <- read_csv('output_data/EVC_2012.csv') %>% # pre-fire
  mutate(severity = factor(severity, levels = c('unb', 'mod', 'high')))
evc_12_meta <- read_csv('raw_data/spatial_data/EVC_2012.csv')
evc_2014 <- read_csv('output_data/EVC_2014.csv') %>% # post-fire
  mutate(severity = factor(severity, levels = c('unb', 'mod', 'high')))
evc_14_meta <- read_csv('raw_data/spatial_data/EVC_2014.csv')

evc_2012_pixels <- read_csv('output_data/EVC_2012_pixels.csv')
evc_2014_pixels <- read_csv('output_data/EVC_2014_pixels.csv')

# vectors

year.labs <- c("2012\nPre-fire", "2014\nPost-fire")
names(year.labs) <- c("2012", "2014")

```

<br>

### Pre-fire similarity

#### Existing Vegetation Cover and Existing Vegetation Type

```{r, include=FALSE}

# Existing Vegetation Cover (EVC)

# 2012 (pre-disturbance)
evc_12 <- evc_2012 %>% 
  group_by(severity) %>%
  mutate(total = sum(n)) %>% 
  group_by(severity, evc) %>% 
  mutate(persev = sum(n),
         prop = persev/total) %>% 
  filter(prop > 0.01) %>% # delete habitats that represent <1% of habitat sampled
  summarize(sum = sum(n)) %>% 
  inner_join(evc_12_meta[,1:2], by = c('evc' = 'VALUE')) # add metadata

# 2014 (post-disturbance)
evc_14 <- evc_2014 %>% 
  group_by(severity) %>%
  mutate(total = sum(n)) %>% 
  group_by(severity, evc) %>% 
  mutate(persev = sum(n),
         prop = persev/total) %>% 
  filter(prop > 0.01) %>% # delete habitats that represent <1% of habitat sampled
  summarize(sum = sum(n)) %>% 
  inner_join(evc_12_meta, by = c('evc' = 'VALUE')) # add metadata

```

```{r, include=FALSE}

year.labs <- c("2012\nPre-fire", "2014\nPost-fire")
names(year.labs) <- c("2012", "2014")

# EVC plot
evc_plot <- full_join(evc_12, evc_14, by = c('severity', 'CLASSNAMES')) %>% 
  rename('2012'=sum.x, '2014'=sum.y, EVC=CLASSNAMES) %>% 
  select(severity, EVC, '2012', '2014') %>% 
  gather(year, pixels, 3:4) %>% 
  ggplot(aes(x = severity, y = pixels, fill = EVC)) +
    geom_bar(position = 'fill', stat = 'identity') +
    scale_fill_manual(name = '% Cover of the dominant lifeform',
                      values = c('#ffce8a','#e89a61',
                                 '#c7e9c0','#a1d99b','#74c476','#41ab5d','#238b45','#005a32'),
                      labels = c('Herb: 20-30%',
                                 'Shrub: 50-60%',
                                 'Tree: 20-30%','Tree: 30-40%','Tree: 40-50%',
                                 'Tree: 50-60%','Tree: 60-70%','Tree: 70-80%')) +
    labs(y = 'Proportional area') +
    scale_y_continuous(expand = c(0,0),
                       labels = c(0,25,50,75,100)) +
    theme_classic() +
    theme(axis.title.x = element_blank(),
          axis.ticks.x = element_blank(),
          axis.title.y = element_text(size = 11),
          axis.text.x = element_blank(),
          axis.text.y = element_text(size = 12, color = 'black'),
          legend.text = element_text(size = 11),
          legend.justification = c(0,0),
          strip.text.x = element_text(size = 12, face = 'bold'),
          plot.margin = unit(c(5,5,5,20), 'pt')) +
  expand_limits(y = 1.1) +
  facet_wrap(~year,
             labeller = labeller(year = year.labs))
    
```

```{r, include=FALSE}

# Existing Vegetation Type (EVT)

# 2012 (pre-disturbance)
evt_12 <- evt_2012 %>% 
  group_by(severity) %>% 
  mutate(total = sum(n)) %>% 
  group_by(severity, evt) %>% 
  mutate(persev = sum(n),
         prop = persev/total) %>% 
  filter(prop > 0.01) %>% # delete habitats that represent <1% of habitat sampled for each severity
  summarize(sum = sum(n)) %>% 
  inner_join(evt_12_meta, by = c('evt' = 'VALUE')) # add metadata

# 2014 (post-disturbance)
evt_14 <- evt_2014 %>% 
  group_by(severity) %>% 
  mutate(total = sum(n)) %>% 
  group_by(severity, evt) %>% 
  mutate(persev = sum(n),
         prop = persev/total) %>% 
  filter(prop > 0.01) %>%
  summarize(sum = sum(n)) %>% 
  inner_join(evt_14_meta, by = c('evt' = 'VALUE'))

```

```{r, include=FALSE}

# EVT plot
evt_plot <- full_join(evt_12, evt_14, by = c('severity', 'CLASSNAME')) %>% 
  rename('2012'=sum.x, '2014'=sum.y, EVT=CLASSNAME) %>% 
  select(severity, EVT, '2012', '2014') %>% 
  mutate(EVT = factor(EVT, levels = c(
    'Mediterranean California Dry-Mesic Mixed Conifer Forest and Woodland',
    'Mediterranean California Mesic Mixed Conifer Forest and Woodland',
    'California Montane Jeffrey Pine(-Ponderosa Pine) Woodland',
    'Mediterranean California Red Fir Forest',
    'California Montane Riparian Systems',
    'California Montane Woodland and Chaparral',
    'North Pacific Montane Grassland'))) %>%
  gather(year, pixels, 3:4) %>% 
  ggplot(aes(x = severity, y = pixels, fill = EVT)) +
    geom_bar(position = 'fill', stat = 'identity') +
    scale_fill_manual(name = 'Vegetation Type',
                      values = c('#9C9C00','#696900','#1DA300','#136700',
                                 '#6baed6',
                                 '#8c2d04','#fe9929'),
                      labels = c('Dry-mesic mixed conifer forest',
                                 'Mesic mixed conifer forest',
                                 'Montane yellow pine woodland',
                                 'Red fir forest',
                                 'Montane riparian system',
                                 'Montane woodland and chaparral',
                                 'Montane grassland')) +
    labs(y = 'EVT (% of pixels)') +
    scale_x_discrete(labels = c('Unburned', 'Low-Mod.\nSeverity', 'High\nSeverity')) +
    scale_y_continuous(expand = c(0,0),
                       labels = c(0,25,50,75,100)) +
    theme_classic() +
    theme(axis.title.x = element_blank(),
          axis.ticks.x = element_blank(),
          axis.title.y = element_text(size = 12),
          axis.text.x = element_text(size = 11, color = 'black', angle = 90, vjust = 0.5),
          axis.text.y = element_text(size = 11, color = 'black'),
          legend.text = element_text(size = 11),
          strip.background = element_blank(),
          strip.text.x = element_blank(),
          legend.justification = c(0,1),
          plot.margin = unit(c(5,5,5,20), 'pt')) +
  expand_limits(y = 1.1) +
  facet_wrap(~year)

```

```{r, fig.height=5.5, fig.width=6}

# Combined plot: EVC, EVT
plot_grid(evc_plot, evt_plot,
          ncol = 1, align = 'v',
          rel_heights = c(1, 1))

```

<br>

#### Existing Vegetation Cover and California Wildlife Habitat Relationships

```{r}

# California Wildlife Habitat Relationships (CWHR)

# Crosswalking from EVT to CWHR data
cwhr_2012 <- left_join(evt_2012, crosswalk)
cwhr_2014 <- left_join(evt_2014, crosswalk)

# 2012 (pre-disturbance)
cwhr_12 <- cwhr_2012 %>% 
  group_by(severity) %>% 
  mutate(total = sum(n)) %>% 
  group_by(severity, cwhr) %>% 
  mutate(persev = sum(n),
         prop = persev/total) %>% 
  filter(prop > 0.01) %>% # delete habitats that represent <1% of habitat sampled for each sev
  summarize(sum = sum(n))

# 2014 (post-disturbance)
cwhr_14 <- cwhr_2014 %>% 
  group_by(severity) %>% 
  mutate(total = sum(n)) %>% 
  group_by(severity, cwhr) %>% 
  mutate(persev = sum(n),
         prop = persev/total) %>% 
  filter(prop > 0.01) %>%
  summarize(sum = sum(n))

```

```{r, include=FALSE}

# CWHR plot
cwhr_plot <- full_join(cwhr_12, cwhr_14, by = c('severity', 'cwhr')) %>% 
  rename('2012' = sum.x, '2014' = sum.y) %>% 
  select(severity, cwhr, '2012', '2014') %>% 
  mutate(cwhr = factor(cwhr, levels = c('Perennial grassland',
                                        'Montane chaparral',
                                        'Montane riparian',
                                        'Ponderosa pine/Jeffrey pine/Eastside pine',
                                        'Red fir',
                                        'Sierran mixed conifer forest/White fir/Douglas fir'))) %>%
  gather(year, pixels, 3:4) %>% 
  ggplot(aes(x = severity, y = pixels, fill = cwhr)) +
    geom_bar(position = 'fill', stat = 'identity') +
    scale_fill_manual(name = 'CWHR habitat type',
                      values = c('#EDDB3B', '#DD9C33', '#6baed6',
                                 '#79BC2D', '#6F8932', '#ACBA2E'),
                      labels = c('Perennial grassland',
                                 'Montane chaparral',
                                 'Montane riparian',
                                 'Ponderosa/Jeffrey pine',
                                 'Red fir',
                                 'Sierran mixed conifer forest')) +
    labs(y = 'Proportional area') +
    scale_x_discrete(labels = c('Unburned', 'Low-mod.\nseverity', 'High\nseverity')) +
    scale_y_continuous(expand = c(0,0),
                       labels = c(0,25,50,75,100)) +
    theme_classic() +
    theme(axis.title.x = element_blank(),
          axis.ticks.x = element_blank(),
          axis.title.y = element_text(size = 12),
          axis.text.x = element_text(size = 11, color = 'black', angle = 90, vjust = 0.5),
          axis.text.y = element_text(size = 11, color = 'black'),
          legend.text = element_text(size = 11),
          strip.background = element_blank(),
          strip.text.x = element_blank(),
          legend.justification = c(0,1),
          plot.margin = unit(c(5,5,5,20), 'pt')) +
  expand_limits(y = 1.1) +
  facet_wrap(~year)

```

```{r, fig.height=5.5, fig.width=6}

# Combined plot: EVC, CWHR
plot_grid(evc_plot, cwhr_plot,
          ncol = 1, align = 'v',
          rel_heights = c(1, 1.1))

ggsave('figs/fig1C.png', plot = last_plot(),
       width = 6, height = 5.5, units = 'in')

```

<br>

#### For each block

- Block 1: Devil's Peak
  - GRNCAR, GRNKAN, GRNWOO, HICOWS, MIXDP2, MIXMID, MIXRAV HISQUE, HIWALK
- Block 2: Belt
  - GRNHOU, GRNMAS, GRNTUR, MIXBUS, MIXMIX, MIXMOR, HICORN, HIDODD, HITEDD
- Block 3: High Meadows
  - GRNDAM, GRNQUA, GRNSNO, MIXBUT, MIXRIV, MIXWAT, HIBADR, HICAMP, HISHRU

```{r}

# Existing Vegetation Cover (EVC)
evc_block <- full_join(evc_2012, site_meta)%>% 
  left_join(evc_12_meta, by = c('evc' = 'VALUE')) %>% 
  mutate(block=factor(block, levels=c('dp','belt','hm'))) %>% 
  group_by(block) %>% 
  mutate(total = sum(n)) %>% 
  group_by(block, CLASSNAMES) %>% 
  mutate(persev = sum(n),
         prop = persev/total) %>% 
  filter(prop > 0.01) %>%
  summarize(sum = sum(n))

# Plot
evc_block %>% 
  ggplot(aes(x = block, y = sum, fill = CLASSNAMES)) +
    geom_bar(position = 'fill', stat = 'identity') +
    scale_fill_manual(name = '% Cover of the Dominant Lifeform',
                      values = c('#e89a61',
                                 '#c7e9c0','#a1d99b','#74c476',
                                 '#41ab5d','#238b45','#005a32'),
                      labels = c('Shrub: 50-60%',
                                 'Tree: 20-30%','Tree: 30-40%','Tree: 40-50%',
                                 'Tree: 50-60%','Tree: 60-70%','Tree: 70-80%')) +
    labs(y = 'Proportional Area') +
    scale_x_discrete(labels = c('1: Devil\'s\nPeak','2: Belt','3: High\nMeadows')) +
    scale_y_continuous(expand = c(0,0),
                       labels = c(0,25,50,75,100)) +
    theme_classic() +
    theme(axis.title.x = element_blank(),
          axis.ticks.x = element_blank(),
          axis.title.y = element_text(size = 12),
          axis.text.x = element_text(size = 11, color = 'black', angle = 90, vjust = 0.5),
          axis.text.y = element_text(size = 11, color = 'black'),
          legend.text = element_text(size = 11),
          legend.justification = c(0,1)) +
  expand_limits(y = 1.1)

```

```{r, fig.height=3, fig.width=8.1}

# Existing Vegetation Type (EVT)
evt_block <- full_join(evt_2012, site_meta)%>% 
  left_join(evt_12_meta, by = c('evt' = 'VALUE')) %>% 
  mutate(block=factor(block, levels=c('dp','belt','hm'))) %>% 
  mutate(CLASSNAME = factor(CLASSNAME, levels = c(
    'Mediterranean California Dry-Mesic Mixed Conifer Forest and Woodland',
    'Mediterranean California Mesic Mixed Conifer Forest and Woodland',
    'California Montane Jeffrey Pine(-Ponderosa Pine) Woodland',
    'Mediterranean California Red Fir Forest',
    'California Montane Riparian Systems',
    'California Montane Woodland and Chaparral'))) %>% 
  group_by(block) %>% 
  mutate(total = sum(n)) %>% 
  group_by(block, CLASSNAME) %>% 
  mutate(persev = sum(n),
         prop = persev/total) %>% 
  filter(prop > 0.01) %>%
  summarize(sum = sum(n))

# Plot
evt_block %>% 
  ggplot(aes(x = block, y = sum, fill = CLASSNAME)) +
    geom_bar(position = 'fill', stat = 'identity') +
    scale_fill_manual(name = 'Existing Vegetation Type (EVT)',
                      values = c('#9C9C00','#696900','#1DA300','#136700',
                                 '#6baed6',
                                 '#8c2d04')) +
    labs(y = 'Proportional Area') +
    scale_x_discrete(labels = c('1: Devil\'s\nPeak','2: Belt','3: High\nMeadows')) +
    scale_y_continuous(expand = c(0,0),
                       labels = c(0,25,50,75,100)) +
    theme_classic() +
    theme(axis.title.x = element_blank(),
          axis.ticks.x = element_blank(),
          axis.title.y = element_text(size = 12),
          axis.text.x = element_text(size = 11, color = 'black', angle = 90, vjust = 0.5),
          axis.text.y = element_text(size = 11, color = 'black'),
          legend.text = element_text(size = 11),
          legend.justification = c(0,1)) +
  expand_limits(y = 1.1)

```

```{r}

# California Wildlife Habitat Relationships (CWHR) - 2012
cwhr_block <- full_join(evt_2012, site_meta) %>% 
  left_join(crosswalk) %>% 
  mutate(block=factor(block, levels=c('dp','belt','hm'))) %>% 
  group_by(block) %>% 
  mutate(total = sum(n)) %>% 
  group_by(block, cwhr) %>% 
  mutate(persev = sum(n),
         prop = persev/total) %>% 
  filter(prop > 0.01) %>%
  summarize(sum = sum(n))

# Plot
cwhr_block %>% 
  mutate(cwhr = factor(cwhr, levels = c('Montane chaparral',
                                        'Montane riparian',
                                        'Ponderosa pine/Jeffrey pine/Eastside pine',
                                        'Red fir',
                                        'Sierran mixed conifer forest/White fir/Douglas fir'))) %>%
  ggplot(aes(x = block, y = sum, fill = cwhr)) +
    geom_bar(position = 'fill', stat = 'identity') +
    scale_fill_manual(name = 'CWHR Habitat Type',
                      values = c('#DD9C33', '#6baed6',
                                 '#79BC2D', '#6F8932', '#ACBA2E'),
                      labels = c('Montane chaparral',
                                 'Montane riparian',
                                 'Ponderosa/Jeffrey pine',
                                 'Red fir',
                                 'Sierran mixed conifer forest')) +
    labs(y = 'Proportional Area') +
    scale_x_discrete(labels = c('1: Devil\'s\nPeak','2: Belt','3: High\nMeadows')) +
    scale_y_continuous(expand = c(0,0),
                       labels = c(0,25,50,75,100)) +
    theme_classic() +
    theme(axis.title.x = element_blank(),
          axis.ticks.x = element_blank(),
          axis.title.y = element_text(size = 12),
          axis.text.x = element_text(size = 11, color = 'black', angle = 90, vjust = 0.5),
          axis.text.y = element_text(size = 11, color = 'black'),
          legend.text = element_text(size = 11),
          legend.justification = c(0,1)) +
  expand_limits(y = 1.1)

```

<br>

#### Stats on EVC

```{r}

# Filtering for tree only
tree_12 <- full_join(evc_2012_pixels, site_meta[,2:4]) %>% 
  left_join(evc_12_meta[,1:2], by = c('evc' = 'VALUE')) %>% 
  mutate(
    class = case_when(
      grepl('Shrub', CLASSNAMES) ~ 'Dominant-other',
      grepl('Herb', CLASSNAMES) ~ 'Dominant-other',
      TRUE ~ CLASSNAMES),
    class = factor(class),
    year = 'prefire') %>% 
  select(-evc, -CLASSNAMES) %>% 
  na.omit(.)
classes <- levels(tree_12$class)
tree_12 <- tree_12 %>% 
  mutate(dummy = case_when(
    class==classes[1] ~ 1,
    class==classes[2] ~ 2,
    class==classes[3] ~ 3,
    class==classes[4] ~ 4,
    class==classes[5] ~ 5,
    class==classes[6] ~ 6,
    class==classes[7] ~ 7,
    class==classes[8] ~ 8
  ))

tree_14 <- full_join(evc_2014_pixels, site_meta[,2:4]) %>% 
  left_join(evc_14_meta[,1:2], by = c('evc' = 'VALUE')) %>% 
  mutate(
    class = case_when(
      grepl('Shrub', CLASSNAMES) ~ 'Dominant-other',
      grepl('Herb', CLASSNAMES) ~ 'Dominant-other',
      TRUE ~ CLASSNAMES),
    class = factor(class),
    year = 'postfire') %>% 
  select(-evc, -CLASSNAMES) %>% 
  na.omit(.)
classes <- levels(tree_14$class)
tree_14 <- tree_14 %>% 
  mutate(dummy = case_when(
    class==classes[1] ~ 1,
    class==classes[2] ~ 2,
    class==classes[3] ~ 3,
    class==classes[4] ~ 4,
    class==classes[5] ~ 5,
    class==classes[6] ~ 6,
    class==classes[7] ~ 7,
    class==classes[8] ~ 8
  ))

tree <- rbind(tree_12, tree_14)

```

**Comparing pre vs post fire**

```{r}

# Check assumptions
ggplot(tree, aes(x=dummy)) +
  geom_histogram() +
  facet_wrap(~severity)

# Wilcoxon ranked sum test

tree_unb <- tree %>% 
  filter(severity=='unb')%>% 
  wilcox_test(dummy ~ year)
tree_mod <- tree %>% 
  filter(severity=='mod') %>% 
  wilcox_test(dummy ~ year)
tree_high <- tree %>% 
  filter(severity=='high') %>% 
  wilcox_test(dummy ~ year)
tree_unb
tree_mod
tree_high

p.adjust(c(tree_unb$p, tree_mod$p, tree_high$p), method='bonferroni')

# unb: 0.573
# mod: 2.673e-36
# high: 9.810e-298

tree %>% filter(severity=='unb') %>% wilcox_effsize(dummy~year)
# effsize 0.0391, n=1120 (small)
tree %>% filter(severity=='mod') %>% wilcox_effsize(dummy~year)
# effsize 0.328, n=1488 (moderate)
tree %>% filter(severity=='high') %>% wilcox_effsize(dummy~year)
# effsize 0.941, n=1537 (large)

```

- Used Wilcoxon ranked sum test because the data is ordinal
- Bonferroni correction for multiple comparisons

**Comparing three severities pre-fire vs. postfire**

```{r}

# Kruskal-Wallis ranked sum test
tree_2012 <- tree_12 %>% 
  kruskal_test(dummy ~ severity)
tree_2014 <- tree_14 %>% 
  kruskal_test(dummy ~ severity)
tree_2012
tree_2014

p.adjust(c(tree_2012$p, tree_2014$p), method='bonferroni')
# 2012: p=0.00424
# 2014: p=0.00000...

tree_12 %>% kruskal_effsize(dummy~severity)
# effsize 0.00564, n=1830 (small)
tree_14 %>% kruskal_effsize(dummy~severity)
# effsize 0.663, n=2315 (large)

dunn_test(tree_14, dummy~severity, p.adjust.method = 'bonferroni')
# all p<0.0000000000000000001

```

- Used Kruskal-Wallis ranked sum test because the data is ordinal
- Bonferroni correction for multiple comparisons
- Post-hoc by Dunn's test with Bonferroni correction for multiple comparisons

**Comparing three blocks pre-fire**

```{r}

# Kruskal-Wallis ranked sum test
tree_block <- tree_12 %>% 
  kruskal_test(dummy ~ block)
tree_block
# p=8.67e-24

tree_12 %>% kruskal_effsize(dummy~block)
# effsize=0.0570, n=1830 (small)

```

- Used Kruskal-Wallis ranked sum test because the data is ordinal

<br>

### Fire categories

**Fire severity metrics for the three severity categories:**

```{r, fig.height=6, fig.width=4}

# Summarize metrics by severity

metrics <- full_join(spatial, veg) %>%
  mutate(severity = factor(severity, levels = c('unb', 'mod', 'high')),
         tree_mortality = case_when(
           severity == 'unb' ~ 0,
           TRUE ~ (1-tree_survival)*100
         ),
         dist_to_edge = dist_to_edge/1000) %>% 
  select(site, severity, avg_pixel, dist_to_edge, tree_mortality)

metrics_summary <- metrics %>% 
  group_by(severity) %>% 
  dplyr::summarize(avg_pixel_mean = mean(avg_pixel),
                   avg_pixel_se = sd(avg_pixel)/sqrt(length(avg_pixel)),
                   dist_to_edge_mean = mean(dist_to_edge),
                   dist_to_edge_se = sd(dist_to_edge)/sqrt(length(dist_to_edge)),
                   tree_mortality_mean = mean(tree_mortality),
                   tree_mortality_se = sd(tree_mortality)/sqrt(length(tree_mortality)))

# Plot metrics by severity

avg_pixel_plot <- ggplot(metrics_summary) +
  geom_col(aes(x = severity, y = avg_pixel_mean, fill = severity)) +
  scale_fill_manual(values = c('darkgreen', 'orange', 'red')) +
  geom_hline(yintercept = 0) +
  geom_errorbar(aes(x = severity,
                    ymin = avg_pixel_mean-avg_pixel_se,
                    ymax = avg_pixel_mean+avg_pixel_se),
                width=.15) +
  labs(y = 'MTBS Severity') +
  theme_classic() +
  theme(axis.title.x = element_blank(),
        legend.position = 'none',
        axis.line.x = element_blank(),
        axis.ticks.x = element_blank(),
        axis.text.x = element_blank(),
        plot.margin = unit(c(5,5,5,20), 'pt'))

tree_mortality_plot <- ggplot(metrics_summary) +
  geom_col(aes(x = severity, y = tree_mortality_mean, fill = severity)) +
  scale_fill_manual(values = c('darkgreen', 'orange', 'red')) +
  geom_errorbar(aes(x = severity,
                    ymin = tree_mortality_mean-tree_mortality_se,
                    ymax = tree_mortality_mean+tree_mortality_se),
                width=.15) +
  labs(y = '% Tree Mortality') +
  scale_x_discrete(labels = c('Unburned', 'Moderate\nSeverity', 'High\nSeverity')) +
  scale_y_continuous(expand = c(0,0)) +
  theme_classic() +
  theme(axis.title.x = element_blank(),
        legend.position = 'none',
        axis.text.x = element_blank(),
        axis.ticks.x = element_blank(),
        plot.margin = unit(c(5,5,5,20), 'pt'))

dist_to_edge_plot <- ggplot(metrics_summary) +
  geom_col(aes(x = severity, y = dist_to_edge_mean, fill = severity)) +
  scale_fill_manual(values = c('darkgreen', 'orange', 'red')) +
  geom_errorbar(aes(x = severity,
                    ymin = dist_to_edge_mean-dist_to_edge_se,
                    ymax = dist_to_edge_mean+dist_to_edge_se),
                width=.15) +
  geom_hline(yintercept = 0) +
  labs(y = 'Dist to edge (km)') +
  scale_x_discrete(labels = c('Unburned', ' Low-Mod\n Severity', 'High\nSeverity')) +
  theme_classic() +
  theme(axis.title.x = element_blank(),
        axis.line.x = element_blank(),
        axis.ticks.x = element_blank(),
        legend.position = 'none',
        plot.margin = unit(c(5,5,5,20), 'pt'))

plot_grid(avg_pixel_plot, tree_mortality_plot, dist_to_edge_plot,
          labels = "AUTO",
          ncol = 1, align = 'v',
          rel_heights = c(1, 1, 1.3))

```

```{r, eval=FALSE}

# Kruskal-Wallis

# Avg pixel
kruskal_test(metrics, avg_pixel ~ severity)
# p<0.0000001
kruskal_effsize(metrics, avg_pixel ~ severity)
# effsize=0.891, n=27 (large)
dunn_test(metrics, avg_pixel ~ severity, p.adjust.method = 'bonferroni')
# unb/mod: p=0.468, unb/high: p=0.00000132, mod/high: p=0.468

# % tree mortality
kruskal_test(metrics, tree_mortality ~ severity)
# p<0.000001
kruskal_effsize(metrics, tree_mortality ~ severity)
# effsize=0.957, n=27 (large)
dunn_test(metrics, tree_mortality ~ severity, p.adjust.method = 'bonferroni')
# unb/mod: p=0.374, unb/high: p=0.00000174, mod/high: p=0.0374

```

- All differences significant for average MTBS severity and % tree mortality (Kruskal Wallis, post-hoc by Dunn's test with Bonferroni correction, $\alpha$ = 0.05)
- Differences in distance-to-edge significant for unburned vs high and unburned vs moderate only

<br>

**Final figure (without dist-to-edge):**

```{r, fig.width=4, fig.height=5}

# Plot metrics by severity WITHOUT dist-to-edge

avg_pixel_plot <- ggplot(metrics_summary) +
  geom_col(aes(x = severity, y = avg_pixel_mean, fill = severity)) +
  scale_fill_manual(values = c('darkgreen', 'orange', 'red')) +
  geom_hline(yintercept = 0) +
  geom_errorbar(aes(x = severity,
                    ymin = avg_pixel_mean-avg_pixel_se,
                    ymax = avg_pixel_mean+avg_pixel_se),
                width=.15) +
  labs(y = 'MTBS Severity') +
  scale_y_continuous(expand = c(0,0)) +
  theme_classic() +
  theme(axis.title.x = element_blank(),
        legend.position = 'none',
        axis.line.x = element_blank(),
        axis.ticks.x = element_blank(),
        axis.text.x = element_blank(),
        axis.title.y = element_text(size = 14),
        axis.text.y = element_text(size = 12, color = 'black'),
        plot.margin = unit(c(10,5,5,20), 'pt'))

tree_mortality_plot <- ggplot(metrics_summary) +
  geom_col(aes(x = severity, y = tree_mortality_mean, fill = severity)) +
  scale_fill_manual(values = c('darkgreen', 'orange', 'red')) +
  geom_errorbar(aes(x = severity,
                    ymin = tree_mortality_mean-tree_mortality_se,
                    ymax = tree_mortality_mean+tree_mortality_se),
                width=.15) +
  labs(y = '% Tree Mortality') +
  scale_x_discrete(labels = c('Unburned', ' Low-Mod.\n Severity', 'High\nSeverity')) +
  scale_y_continuous(expand = c(0,0)) +
  theme_classic() +
  theme(legend.position = 'none',
        axis.title.x = element_blank(),
        axis.ticks.x = element_blank(),
        axis.title.y = element_text(size = 14),
        axis.text.x = element_text(size = 10, color = 'black'),
        axis.text.y = element_text(size = 12, color = 'black'),
        plot.margin = unit(c(10,5,5,20), 'pt'))

plot_grid(avg_pixel_plot, tree_mortality_plot,
          ncol = 1, align = 'v',
          rel_heights = c(1, 1))

ggsave('figs/fig1D.png', plot = last_plot(),
       width = 3, height = 4, units = 'in')

```

```{r}

colors <- c('#4AB793', '#DCA827', '#D85F2B')
sevs <- c('Unburned', 'Low-mod.\nseverity', 'High\nseverity')

set.seed(3)
mtbs_plot <- metrics %>% 
  ggplot() +
  geom_jitter(aes(x=severity, y=avg_pixel, color=severity, shape=severity),
              width=0.1, size=1.5) +
  geom_boxplot(aes(x=severity, y=avg_pixel), fill=NA, outlier.shape=NA) +
  scale_color_manual(values=colors) +
  scale_shape_manual(values=c(19,17,15)) +
  theme_classic() +
  theme(legend.position = 'NA',
        axis.text.y = element_text(size = 9),
        axis.text.x = element_blank()) +
  scale_y_continuous(limits=c(-0.01,4.5), breaks=c(0,1,2,3,4)) +
  labs(x='', y='MTBS severity')

set.seed(4)
tree_plot <- metrics %>% 
  ggplot() +
  geom_jitter(aes(x=severity, y=tree_mortality, color=severity, shape=severity),
              width=0.1, size=1.5) +
  geom_boxplot(aes(x=severity, y=tree_mortality), fill=NA, outlier.shape=NA) +
  scale_color_manual(values=colors) +
  scale_shape_manual(values=c(19,17,15)) +
  theme_classic() +
  theme(legend.position = 'NA',
        axis.text.y = element_text(size = 9),
        axis.text.x = element_text(size = 7)) +
  scale_y_continuous(limits=c(-0.2,100.2), breaks=c(0,25,50,75,100)) +
  scale_x_discrete(labels=sevs) +
  labs(x='', y='% Tree mortality')

plot_grid(mtbs_plot, tree_plot,
          ncol = 1, align = 'v',
          rel_heights = c(1, 1, 1.1))

ggsave('figs/fig1D.png', plot = last_plot(),
       width = 2, height = 3, units = 'in')

```

<br>

### PCA of the three severity metrics

```{r, fig.height=5}

# PCA of the three severity metrics

severity_pca <- prcomp(metrics[c(3:5)],
                       scale = TRUE)

ggbiplot(severity_pca, groups = metrics$severity) + # PC1 describes 88% of the variance
  scale_color_manual(values = c('darkgreen', 'orange', 'red')) +
  theme_classic() +
  theme(legend.position = 'none')

```

<br>

**Number lines showing the 27 sites by severity metrics:**

```{r, fig.height=1}

# Number lines of the severity groups

severity_scores <- data.frame(severity_pca$x) %>% 
  bind_cols(metrics) %>% 
  select(site, severity, severity_score = PC1)

ggplot(severity_scores, aes(x = severity_score, y = 1, color = severity)) +
  geom_point() +
  scale_color_manual(values = c('darkgreen', 'orange', 'red')) +
  labs(x = 'Severity PCA Score', y = '') +
  theme_void() +
  theme(axis.title = element_text(),
        axis.line.x = element_line(),
        axis.text.x = element_text())

ggplot(metrics, aes(x = avg_pixel, y = 1, color = severity)) +
  geom_point() +
  scale_color_manual(values = c('darkgreen', 'orange', 'red')) +
  labs(x = 'Average MTBS Pixel', y = '') +
  theme_void() +
  theme(axis.title = element_text(),
        axis.line.x = element_line(),
        axis.text.x = element_text())

```

```{r eval=FALSE}
rmarkdown::render('code/12_site_justification.Rmd', output_file = '../docs/12_site_justification.html')
```
