==> code/13_summary_fig.Rmd <==
---
title: "Summary Figure"
author: "Kate Culhane"
output: html_document
editor_options: 
  chunk_output_type: console
---

```{r setup, include=FALSE}
knitr::opts_knit$set(root.dir = '..')
knitr::opts_chunk$set(echo = FALSE, warning = FALSE, message = FALSE,
                      fig.width=5, fig.height=3)
```

<br>

```{r packages & data}

# Load packages

library(tidyverse)
library(ggnewscale)
library(scales)
library(ggthemes)
library(cowplot) # plot_grid function

# Load data

abundance <- read_csv('output_data/abundance.csv')
diversity <- read_csv('output_data/diversity.csv')
FD <- read_csv('output_data/FD_biomass.csv')
PD <- read_csv('output_data/PD.csv')

colors <- c('#4AB793', '#DCA827', '#D85F2B')
sevs <- c('Unburned', 'Low-mod.\nseverity', 'High\nseverity')

```

```{r}

# Combined data frame
metrics <- full_join(abundance, FD) %>% 
  full_join(diversity) %>% 
  full_join(PD) %>% 
  mutate(severity = factor(severity, levels = c('unb', 'mod', 'high')),
         biomass = biomass_total/1000) %>% 
    select(site, severity,
           n_adj, biomass, n_adj_pema, n_adj_sotr,
           rarefied_richness, J, FEve, PD)

# Summary data frame 
summary <- metrics%>% 
  gather(key = 'metric', value = 'value', 3:10) %>% 
  group_by(metric, severity) %>% 
  summarize(mean = mean(value, na.rm = TRUE),
            se = sd(value, na.rm = TRUE)/sqrt(length(value))) %>% 
  ungroup() %>% 
  mutate(metric = factor(metric, levels = c('n_adj', 'biomass',
                                            'n_adj_pema', 'n_adj_sotr',
                                            'rarefied_richness', 'J',
                                            'FEve', 'PD')))

# Vectors
metric_labs <- c('Total abundance\n(indv/site)', 'Biomass\n(kg/site)',
              'Deer mouse\nabundance (indv/site)', 'Trowbridge\'s shrew\nabundance (indv/site)',
              'Rarefied\nrichness', 'Evenness',
              'FEve', 'Faith\'s PD')
names(metric_labs) <- c('n_adj', 'biomass', 'n_adj_pema', 'n_adj_sotr',
                        'rarefied_richness', 'J', 'FEve', 'PD')

```

### Summary Figure

```{r}

set.seed(2)
abun_plot <- metrics %>% 
  ggplot() +
  geom_jitter(aes(x=severity, y=n_adj, color=severity, shape=severity), width=0.1, size=1.5) +
  geom_boxplot(aes(x=severity, y=n_adj), fill=NA, outlier.shape=NA) +
  scale_color_manual(values=colors) +
  scale_shape_manual(values=c(19,17,15)) +
  theme_classic() +
  theme(legend.position = 'NA',
        axis.text.y = element_text(size = 9),
        axis.text.x = element_blank()) +
  scale_y_continuous(limits=c(0,60), breaks=c(0,20,40,60)) +
  labs(x='', y='Total abundance\n(indv/site)')

set.seed(5)
biomass_plot <- metrics %>% 
  ggplot() +
  geom_jitter(aes(x=severity, y=biomass, color=severity, shape=severity), width=0.1, size=1.5) +
  geom_boxplot(aes(x=severity, y=biomass), fill=NA, outlier.shape=NA) +
  scale_color_manual(values=colors) +
  scale_shape_manual(values=c(19,17,15)) +
  theme_classic() +
  theme(legend.position = 'NA',
        axis.text.y = element_text(size = 9),
        axis.text.x = element_blank()) +
  scale_y_continuous(limits=c(0,3.1), breaks=c(0,1,2,3)) +
  labs(x='', y='Biomass\n(kg/site)')

set.seed(2)
pema_plot <- metrics %>% 
  ggplot() +
  geom_jitter(aes(x=severity, y=n_adj_pema, color=severity, shape=severity),
              width=0.1, size=1.5) +
  geom_boxplot(aes(x=severity, y=n_adj_pema), fill=NA, outlier.shape=NA) +
  scale_color_manual(values=colors) +
  scale_shape_manual(values=c(19,17,15)) +
  theme_classic() +
  theme(legend.position = 'NA',
        axis.text.y = element_text(size = 9),
        axis.text.x = element_blank()) +
  scale_y_continuous(limits=c(0,60), breaks=c(0,20,40,60)) +
  labs(x='', y='Deer mouse\nabundance (indv/site)')

set.seed(13)
sotr_plot <- metrics %>% 
  ggplot() +
  geom_jitter(aes(x=severity, y=n_adj_sotr, color=severity, shape=severity),
              width=0.1, size=1.5) +
  geom_boxplot(aes(x=severity, y=n_adj_sotr), fill=NA, outlier.shape=NA) +
  scale_color_manual(values=colors) +
  scale_shape_manual(values=c(19,17,15)) +
  theme_classic() +
  theme(legend.position = 'NA',
        axis.text.y = element_text(size = 9),
        axis.text.x = element_blank()) +
  scale_y_continuous(limits=c(0,7), breaks=c(0,2,4,6)) +
  labs(x='', y='Trowbridge\'s shrew\nabundance (indv/site)')

set.seed(24)
rich_plot <- metrics %>% 
  ggplot() +
  geom_jitter(aes(x=severity, y=rarefied_richness, color=severity, shape=severity),
              width=0.1, size=1.5) +
  geom_boxplot(aes(x=severity, y=rarefied_richness), fill=NA, outlier.shape=NA) +
  scale_color_manual(values=colors) +
  scale_shape_manual(values=c(19,17,15)) +
  theme_classic() +
  theme(legend.position = 'NA',
        axis.text.y = element_text(size = 9)) +
  scale_y_continuous(limits=c(0,3.3), breaks=c(0,1,2,3)) +
  scale_x_discrete(labels=sevs) +
  labs(x='', y='Rarefied richness')

set.seed(6)
even_plot <- metrics %>% 
  ggplot() +
  geom_jitter(aes(x=severity, y=J, color=severity, shape=severity),
              width=0.1, size=1.5, na.rm=TRUE) +
  geom_boxplot(aes(x=severity, y=J), fill=NA, outlier.shape=NA, na.rm=TRUE) +
  scale_color_manual(values=colors) +
  scale_shape_manual(values=c(19,17,15)) +
  theme_classic() +
  theme(legend.position = 'NA',
        axis.text.y = element_text(size = 9)) +
  scale_y_continuous(limits=c(0,1.1), breaks=c(0,1,2,3)) +
  scale_x_discrete(labels=sevs) +
  labs(x='', y='Pielou\'s evenness')

plot_grid(abun_plot, biomass_plot, pema_plot, sotr_plot, rich_plot, even_plot,
          ncol = 2, align = 'v',
          rel_heights = c(1, 1, 1.2))

ggsave('figs/fig2.png', plot = last_plot(),
       width = 5, height = 6, units = 'in')
```

```{r eval=FALSE}
rmarkdown::render('code/13_summary_fig.Rmd', output_file = '../docs/13_summary_fig.html')
```
